remove amiibo_data
