--- fw/application/src/amiibo_data.c
@@ -1,843 +1,0 @@
-#include "amiibo_data.h"
-const amiibo_data_t amiibo_data[] = {
-{0x00000000, 0x00000002, "Super Smash Bros.", "Super Mario", "Mario", "Mario", ""}, 
-{0x00000000, 0x00340102, "Super Mario Bros.", "Super Mario", "Mario", "Mario", ""}, 
-{0x00000000, 0x003c0102, "Super Mario Bros.", "Super Mario", "Mario", "Mario - Gold Edition", ""}, 
-{0x00000000, 0x003d0102, "Super Mario Bros.", "Super Mario", "Mario", "Mario - Silver Edition", ""}, 
-{0x00000000, 0x02380602, "8-bit Mario", "Super Mario", "Mario", "8-Bit Mario Classic Color", ""}, 
-{0x00000000, 0x02390602, "8-bit Mario", "Super Mario", "Mario", "8-Bit Mario Modern Color", ""}, 
-{0x00000000, 0x03710102, "Super Mario Bros.", "Super Mario", "Mario", "Mario - Wedding", ""}, 
-{0x00000003, 0x039bff02, "Super Nintendo World", "Super Mario", "Mario", "Mario - Power Up Band", ""}, 
-{0x00000100, 0x00190002, "Super Smash Bros.", "Super Mario", "Mario", "Dr. Mario", ""}, 
-{0x00000300, 0x03a60102, "Super Mario Bros.", "Super Mario", "Mario", "Mario - Cat", ""}, 
-{0x00010000, 0x000c0002, "Super Smash Bros.", "Super Mario", "Luigi", "Luigi", ""}, 
-{0x00010000, 0x00350102, "Super Mario Bros.", "Super Mario", "Luigi", "Luigi", ""}, 
-{0x00010003, 0x039cff02, "Super Nintendo World", "Super Mario", "Luigi", "Luigi - Power Up Band", ""}, 
-{0x00020000, 0x00010002, "Super Smash Bros.", "Super Mario", "Peach", "Peach", ""}, 
-{0x00020000, 0x00360102, "Super Mario Bros.", "Super Mario", "Peach", "Peach", ""}, 
-{0x00020000, 0x03720102, "Super Mario Bros.", "Super Mario", "Peach", "Peach - Wedding", ""}, 
-{0x00020003, 0x039dff02, "Super Nintendo World", "Super Mario", "Peach", "Peach - Power Up Band", ""}, 
-{0x00020100, 0x03a70102, "Super Mario Bros.", "Super Mario", "Peach", "Peach - Cat", ""}, 
-{0x00030000, 0x00020002, "Super Smash Bros.", "Super Mario", "Yoshi", "Yoshi", ""}, 
-{0x00030000, 0x00370102, "Super Mario Bros.", "Super Mario", "Yoshi", "Yoshi", ""}, 
-{0x00030003, 0x039fff02, "Super Nintendo World", "Super Mario", "Yoshi", "Yoshi - Power Up Band", ""}, 
-{0x00030102, 0x00410302, "Yoshi's Woolly World", "Super Mario", "Yoshi", "Green Yarn Yoshi", ""}, 
-{0x00030102, 0x00420302, "Yoshi's Woolly World", "Super Mario", "Yoshi", "Pink Yarn Yoshi", ""}, 
-{0x00030102, 0x00430302, "Yoshi's Woolly World", "Super Mario", "Yoshi", "Light Blue Yarn Yoshi", ""}, 
-{0x00030102, 0x023e0302, "Yoshi's Woolly World", "Super Mario", "Yoshi", "Mega Yarn Yoshi", ""}, 
-{0x00040000, 0x02620102, "Super Mario Bros.", "Super Mario", "Rosalina", "Rosalina", ""}, 
-{0x00040100, 0x00130002, "Super Smash Bros.", "Super Mario", "Rosalina", "Rosalina & Luma", ""}, 
-{0x00050000, 0x00140002, "Super Smash Bros.", "Super Mario", "Bowser", "Bowser", ""}, 
-{0x00050000, 0x00390102, "Super Mario Bros.", "Super Mario", "Bowser", "Bowser", ""}, 
-{0x00050000, 0x03730102, "Super Mario Bros.", "Super Mario", "Bowser", "Bowser - Wedding", ""}, 
-{0x0005ff00, 0x023a0702, "Skylanders", "Super Mario", "Bowser", "Hammer Slam Bowser", ""}, 
-{0x00060000, 0x00150002, "Super Smash Bros.", "Super Mario", "Bowser Jr.", "Bowser Jr.", ""}, 
-{0x00070000, 0x001a0002, "Super Smash Bros.", "Super Mario", "Wario", "Wario", ""}, 
-{0x00070000, 0x02630102, "Super Mario Bros.", "Super Mario", "Wario", "Wario", ""}, 
-{0x00080000, 0x00030002, "Super Smash Bros.", "Super Mario", "Donkey Kong", "Donkey Kong", ""}, 
-{0x00080000, 0x02640102, "Super Mario Bros.", "Super Mario", "Donkey Kong", "Donkey Kong", ""}, 
-{0x0008ff00, 0x023b0702, "Skylanders", "Super Mario", "Donkey Kong", "Turbo Charge Donkey Kong", ""}, 
-{0x00090000, 0x000d0002, "Super Smash Bros.", "Super Mario", "Diddy Kong", "Diddy Kong", ""}, 
-{0x00090000, 0x02650102, "Super Mario Bros.", "Super Mario", "Diddy Kong", "Diddy Kong", ""}, 
-{0x000a0000, 0x00380102, "Super Mario Bros.", "Super Mario", "Toad", "Toad", ""}, 
-{0x000a0003, 0x03a0ff02, "Super Nintendo World", "Super Mario", "Toad", "Toad - Power Up Band", ""}, 
-{0x00130000, 0x02660102, "Super Mario Bros.", "Super Mario", "Daisy", "Daisy", ""}, 
-{0x00130000, 0x037a0002, "Super Smash Bros.", "Super Mario", "Daisy", "Daisy", ""}, 
-{0x00130003, 0x039eff02, "Super Nintendo World", "Super Mario", "Daisy", "Daisy - Power Up Band", ""}, 
-{0x00140000, 0x02670102, "Super Mario Bros.", "Super Mario", "Waluigi", "Waluigi", ""}, 
-{0x00150000, 0x03670102, "Super Mario Bros.", "Super Mario", "Goomba", "Goomba", ""}, 
-{0x00170000, 0x02680102, "Super Mario Bros.", "Super Mario", "Boo", "Boo", ""}, 
-{0x00230000, 0x03680102, "Super Mario Bros.", "Super Mario", "Koopa Troopa", "Koopa Troopa", ""}, 
-{0x00240000, 0x038d0002, "Super Smash Bros.", "Super Mario", "Piranha Plant", "Piranha Plant", ""}, 
-{0x00800102, 0x035d0302, "Yoshi's Woolly World", "Yoshi's Woolly World", "Poochy", "Poochy", ""}, 
-{0x00c00000, 0x037b0002, "Super Smash Bros.", "Donkey Kong", "King K. Rool", "King K. Rool", ""}, 
-{0x01000000, 0x00040002, "Super Smash Bros.", "The Legend of Zelda", "Link", "Link", ""}, 
-{0x01000000, 0x034b0902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Ocarina of Time", ""}, 
-{0x01000000, 0x034c0902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Majora's Mask", ""}, 
-{0x01000000, 0x034d0902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Twilight Princess", ""}, 
-{0x01000000, 0x034e0902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Skyward Sword", ""}, 
-{0x01000000, 0x034f0902, "Legend Of Zelda", "The Legend of Zelda", "Link", "8-Bit Link", ""}, 
-{0x01000000, 0x03530902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Archer", ""}, 
-{0x01000000, 0x03540902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Rider", "旅人缰绳 旅人马鞍"}, 
-{0x01000000, 0x037c0002, "Super Smash Bros.", "The Legend of Zelda", "Link", "Young Link", ""}, 
-{0x01000000, 0x03990902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Link's Awakening", ""}, 
-{0x01000100, 0x00160002, "Super Smash Bros.", "The Legend of Zelda", "Link", "Toon Link", ""}, 
-{0x01000100, 0x03500902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Toon Link - The Wind Waker", ""}, 
-{0x01010000, 0x000e0002, "Super Smash Bros.", "The Legend of Zelda", "Zelda", "Zelda", "黄昏弓"}, 
-{0x01010000, 0x03520902, "Legend Of Zelda", "The Legend of Zelda", "Zelda", "Toon Zelda - The Wind Waker", ""}, 
-{0x01010000, 0x03560902, "Legend Of Zelda", "The Legend of Zelda", "Zelda", "Zelda", ""}, 
-{0x01010100, 0x00170002, "Super Smash Bros.", "The Legend of Zelda", "Zelda", "Sheik", "希克面罩"}, 
-{0x01010300, 0x04140902, "Legend Of Zelda", "The Legend of Zelda", "Zelda", "Zelda & Loftwing", ""}, 
-{0x01020100, 0x001b0002, "Super Smash Bros.", "The Legend of Zelda", "Ganon", "Ganondorf", "六贤者之剑"}, 
-{0x01030000, 0x024f0902, "Legend Of Zelda", "The Legend of Zelda", "Midna", "Midna & Wolf Link", "林克狼"}, 
-{0x01050000, 0x03580902, "Legend Of Zelda", "The Legend of Zelda", "Daruk", "Daruk", "神兽兵装·鲁达尼亚"}, 
-{0x01060000, 0x03590902, "Legend Of Zelda", "The Legend of Zelda", "Urbosa", "Urbosa", "神兽兵装·娜波力斯"}, 
-{0x01070000, 0x035a0902, "Legend Of Zelda", "The Legend of Zelda", "Mipha", "Mipha", "神兽兵装·娜波力斯 卓拉之枪"}, 
-{0x01080000, 0x035b0902, "Legend Of Zelda", "The Legend of Zelda", "Revali", "Revali", "神兽兵装·梅德"}, 
-{0x01400000, 0x03550902, "Legend Of Zelda", "Breath of the Wild", "Guardian", "Guardian", ""}, 
-{0x01410000, 0x035c0902, "Legend Of Zelda", "Breath of the Wild", "Bokoblin", "Bokoblin", ""}, 
-{0x01800000, 0x00080002, "Super Smash Bros.", "Animal Crossing", "Villager", "Villager", ""}, 
-{0x01810000, 0x024b0502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Summer Outfit", ""}, 
-{0x01810000, 0x037d0002, "Super Smash Bros.", "Animal Crossing", "Isabelle", "Isabelle", ""}, 
-{0x01810001, 0x00440502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle", ""}, 
-{0x01810001, 0x01d40502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Character Parfait", ""}, 
-{0x01810100, 0x023f0502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Winter Outfit", ""}, 
-{0x01810101, 0x00b40502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Winter", ""}, 
-{0x01810201, 0x011a0502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Kimono", ""}, 
-{0x01810301, 0x01700502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Dress", ""}, 
-{0x01810401, 0x03aa0502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle", ""}, 
-{0x01810501, 0x03bf0502, "Animal Crossing", "Animal Crossing", "Isabelle", "Isabelle - Sweater", ""}, 
-{0x01820000, 0x02400502, "Animal Crossing", "Animal Crossing", "K.K. Slider", "K. K. Slider", ""}, 
-{0x01820001, 0x00a80502, "Animal Crossing", "Animal Crossing", "K.K. Slider", "K.K. Slider", ""}, 
-{0x01820001, 0x01d80502, "Animal Crossing", "Animal Crossing", "K.K. Slider", "K. K. Slider - Pikopuri", ""}, 
-{0x01820001, 0x03b20502, "Animal Crossing", "Animal Crossing", "K.K. Slider", "K.K. Slider", ""}, 
-{0x01820101, 0x00460502, "Animal Crossing", "Animal Crossing", "K.K. Slider", "DJ KK", ""}, 
-{0x01830000, 0x02420502, "Animal Crossing", "Animal Crossing", "Tom Nook", "Tom Nook", ""}, 
-{0x01830001, 0x00450502, "Animal Crossing", "Animal Crossing", "Tom Nook", "Tom Nook", ""}, 
-{0x01830101, 0x010e0502, "Animal Crossing", "Animal Crossing", "Tom Nook", "Tom Nook - Jacket", ""}, 
-{0x01830201, 0x03a80502, "Animal Crossing", "Animal Crossing", "Tom Nook", "Tom Nook", ""}, 
-{0x01830301, 0x03be0502, "Animal Crossing", "Animal Crossing", "Tom Nook", "Tom Nook - Coat", ""}, 
-{0x01840000, 0x024d0502, "Animal Crossing", "Animal Crossing", "Timmy & Tommy", "Timmy & Tommy", ""}, 
-{0x01840501, 0x03a90502, "Animal Crossing", "Animal Crossing", "Timmy & Tommy", "Timmy & Tommy", ""}, 
-{0x01850001, 0x004b0502, "Animal Crossing", "Animal Crossing", "Timmy", "Timmy", ""}, 
-{0x01850201, 0x01170502, "Animal Crossing", "Animal Crossing", "Timmy", "Timmy - Full Apron", ""}, 
-{0x01850401, 0x01790502, "Animal Crossing", "Animal Crossing", "Timmy", "Timmy - Suit", ""}, 
-{0x01860101, 0x00af0502, "Animal Crossing", "Animal Crossing", "Tommy", "Tommy - Uniform", ""}, 
-{0x01860301, 0x01750502, "Animal Crossing", "Animal Crossing", "Tommy", "Tommy - Suit", ""}, 
-{0x01870001, 0x00470502, "Animal Crossing", "Animal Crossing", "Sable", "Sable", ""}, 
-{0x01870001, 0x03b00502, "Animal Crossing", "Animal Crossing", "Sable", "Sable", ""}, 
-{0x01880000, 0x02410502, "Animal Crossing", "Animal Crossing", "Mabel", "Mabel", ""}, 
-{0x01880001, 0x01120502, "Animal Crossing", "Animal Crossing", "Mabel", "Mabel", ""}, 
-{0x01880001, 0x03af0502, "Animal Crossing", "Animal Crossing", "Mabel", "Mabel", ""}, 
-{0x01890001, 0x00ab0502, "Animal Crossing", "Animal Crossing", "Label", "Label", ""}, 
-{0x01890101, 0x03b10502, "Animal Crossing", "Animal Crossing", "Label", "Label", ""}, 
-{0x018a0000, 0x02450502, "Animal Crossing", "Animal Crossing", "Reese", "Reese", ""}, 
-{0x018a0001, 0x00a90502, "Animal Crossing", "Animal Crossing", "Reese", "Reese", ""}, 
-{0x018b0000, 0x02460502, "Animal Crossing", "Animal Crossing", "Cyrus", "Cyrus", ""}, 
-{0x018b0001, 0x01150502, "Animal Crossing", "Animal Crossing", "Cyrus", "Cyrus", ""}, 
-{0x018c0000, 0x02430502, "Animal Crossing", "Animal Crossing", "Digby", "Digby", ""}, 
-{0x018c0001, 0x004c0502, "Animal Crossing", "Animal Crossing", "Digby", "Digby", ""}, 
-{0x018c0101, 0x01180502, "Animal Crossing", "Animal Crossing", "Digby", "Digby - Raincoat", ""}, 
-{0x018d0000, 0x024c0502, "Animal Crossing", "Animal Crossing", "Rover", "Rover", ""}, 
-{0x018d0001, 0x010c0502, "Animal Crossing", "Animal Crossing", "Rover", "Rover", ""}, 
-{0x018e0000, 0x02490502, "Animal Crossing", "Animal Crossing", "Resetti", "Resetti", ""}, 
-{0x018e0001, 0x00490502, "Animal Crossing", "Animal Crossing", "Resetti", "Resetti", ""}, 
-{0x018e0101, 0x01780502, "Animal Crossing", "Animal Crossing", "Resetti", "Resetti - Without Hat", ""}, 
-{0x018f0001, 0x00b30502, "Animal Crossing", "Animal Crossing", "Don Resetti", "Don Resetti", ""}, 
-{0x018f0101, 0x01190502, "Animal Crossing", "Animal Crossing", "Don Resetti", "Don Resetti - Without Hat", ""}, 
-{0x01900001, 0x01710502, "Animal Crossing", "Animal Crossing", "Brewster", "Brewster", ""}, 
-{0x01910001, 0x004e0502, "Animal Crossing", "Animal Crossing", "Harriet", "Harriet", ""}, 
-{0x01920000, 0x02470502, "Animal Crossing", "Animal Crossing", "Blathers", "Blathers", ""}, 
-{0x01920001, 0x010d0502, "Animal Crossing", "Animal Crossing", "Blathers", "Blathers", ""}, 
-{0x01920001, 0x03ad0502, "Animal Crossing", "Animal Crossing", "Blathers", "Blathers", ""}, 
-{0x01930000, 0x02480502, "Animal Crossing", "Animal Crossing", "Celeste", "Celeste", ""}, 
-{0x01930001, 0x01740502, "Animal Crossing", "Animal Crossing", "Celeste", "Celeste", ""}, 
-{0x01930001, 0x03ae0502, "Animal Crossing", "Animal Crossing", "Celeste", "Celeste", ""}, 
-{0x01940000, 0x024a0502, "Animal Crossing", "Animal Crossing", "Kicks", "Kicks", ""}, 
-{0x01940001, 0x00aa0502, "Animal Crossing", "Animal Crossing", "Kicks", "Kicks", ""}, 
-{0x01940001, 0x03b60502, "Animal Crossing", "Animal Crossing", "Kicks", "Kicks", ""}, 
-{0x01950001, 0x00b00502, "Animal Crossing", "Animal Crossing", "Porter", "Porter", ""}, 
-{0x01960000, 0x024e0502, "Animal Crossing", "Animal Crossing", "Kapp'n", "Kapp'n", ""}, 
-{0x01960001, 0x00480502, "Animal Crossing", "Animal Crossing", "Kapp'n", "Kapp'n", ""}, 
-{0x01970001, 0x01770502, "Animal Crossing", "Animal Crossing", "Leilani", "Leilani", ""}, 
-{0x01980001, 0x00b10502, "Animal Crossing", "Animal Crossing", "Leila", "Leila", ""}, 
-{0x01990001, 0x01160502, "Animal Crossing", "Animal Crossing", "Grams", "Grams", ""}, 
-{0x019a0001, 0x00b70502, "Animal Crossing", "Animal Crossing", "Chip", "Chip", ""}, 
-{0x019b0001, 0x00b60502, "Animal Crossing", "Animal Crossing", "Nat", "Nat", ""}, 
-{0x019c0001, 0x01730502, "Animal Crossing", "Animal Crossing", "Phineas", "Phineas", ""}, 
-{0x019d0001, 0x00ac0502, "Animal Crossing", "Animal Crossing", "Copper", "Copper", ""}, 
-{0x019e0001, 0x00ad0502, "Animal Crossing", "Animal Crossing", "Booker", "Booker", ""}, 
-{0x019f0001, 0x01110502, "Animal Crossing", "Animal Crossing", "Pete", "Pete", ""}, 
-{0x01a00001, 0x010f0502, "Animal Crossing", "Animal Crossing", "Pelly", "Pelly", ""}, 
-{0x01a10001, 0x01100502, "Animal Crossing", "Animal Crossing", "Phyllis", "Phyllis", ""}, 
-{0x01a20001, 0x017d0502, "Animal Crossing", "Animal Crossing", "Gulliver", "Gulliver", ""}, 
-{0x01a20001, 0x03b90502, "Animal Crossing", "Animal Crossing", "Gulliver", "Gulliver", ""}, 
-{0x01a30001, 0x004a0502, "Animal Crossing", "Animal Crossing", "Joan", "Joan", ""}, 
-{0x01a40001, 0x004d0502, "Animal Crossing", "Animal Crossing", "Pascal", "Pascal", ""}, 
-{0x01a50001, 0x01720502, "Animal Crossing", "Animal Crossing", "Katrina", "Katrina", ""}, 
-{0x01a60001, 0x00500502, "Animal Crossing", "Animal Crossing", "Saharah", "Saharah", ""}, 
-{0x01a60001, 0x03b70502, "Animal Crossing", "Animal Crossing", "Saharah", "Saharah", ""}, 
-{0x01a70001, 0x01140502, "Animal Crossing", "Animal Crossing", "Wendell", "Wendell", ""}, 
-{0x01a80001, 0x004f0502, "Animal Crossing", "Animal Crossing", "Redd", "Redd", ""}, 
-{0x01a80101, 0x017e0502, "Animal Crossing", "Animal Crossing", "Redd", "Redd - Shirt", ""}, 
-{0x01a90001, 0x01760502, "Animal Crossing", "Animal Crossing", "Gracie", "Gracie", ""}, 
-{0x01aa0001, 0x00530502, "Animal Crossing", "Animal Crossing", "Lyle", "Lyle", ""}, 
-{0x01ab0001, 0x017c0502, "Animal Crossing", "Animal Crossing", "Pave", "Pave", ""}, 
-{0x01ac0001, 0x017f0502, "Animal Crossing", "Animal Crossing", "Zipper", "Zipper", ""}, 
-{0x01ad0001, 0x00b80502, "Animal Crossing", "Animal Crossing", "Jack", "Jack", ""}, 
-{0x01ae0001, 0x011b0502, "Animal Crossing", "Animal Crossing", "Franklin", "Franklin", ""}, 
-{0x01af0001, 0x011c0502, "Animal Crossing", "Animal Crossing", "Jingle", "Jingle", ""}, 
-{0x01b00001, 0x00520502, "Animal Crossing", "Animal Crossing", "Tortimer", "Tortimer", ""}, 
-{0x01b10001, 0x00b20502, "Animal Crossing", "Animal Crossing", "Shrunk", "Shrunk", ""}, 
-{0x01b10101, 0x017b0502, "Animal Crossing", "Animal Crossing", "Shrunk", "Shrunk - Loud Jacket", ""}, 
-{0x01b30001, 0x00b50502, "Animal Crossing", "Animal Crossing", "Blanca", "Blanca", ""}, 
-{0x01b40001, 0x01130502, "Animal Crossing", "Animal Crossing", "Leif", "Leif", ""}, 
-{0x01b50001, 0x00510502, "Animal Crossing", "Animal Crossing", "Luna", "Luna", ""}, 
-{0x01b60001, 0x00ae0502, "Animal Crossing", "Animal Crossing", "Katie", "Katie", ""}, 
-{0x01c10000, 0x02440502, "Animal Crossing", "Animal Crossing", "Lottie", "Lottie", ""}, 
-{0x01c10001, 0x00540502, "Animal Crossing", "Animal Crossing", "Lottie", "Lottie", ""}, 
-{0x01c10101, 0x017a0502, "Animal Crossing", "Animal Crossing", "Lottie", "Lottie - Black Skirt And Bow", ""}, 
-{0x01c10201, 0x03bb0502, "Animal Crossing", "Animal Crossing", "Lottie", "Lottie - Island", ""}, 
-{0x02000001, 0x00a10502, "Animal Crossing", "Animal Crossing", "Cyrano", "Cyrano", ""}, 
-{0x02010001, 0x016a0502, "Animal Crossing", "Animal Crossing", "Antonio", "Antonio", ""}, 
-{0x02020001, 0x01030502, "Animal Crossing", "Animal Crossing", "Pango", "Pango", ""}, 
-{0x02030001, 0x019a0502, "Animal Crossing", "Animal Crossing", "Anabelle", "Anabelle", ""}, 
-{0x02060001, 0x03120502, "Animal Crossing", "Animal Crossing", "Snooty", "Snooty", ""}, 
-{0x02080001, 0x00960502, "Animal Crossing", "Animal Crossing", "Annalisa", "Annalisa", ""}, 
-{0x02090001, 0x019f0502, "Animal Crossing", "Animal Crossing", "Olaf", "Olaf", ""}, 
-{0x02140001, 0x00e40502, "Animal Crossing", "Animal Crossing", "Teddy", "Teddy", ""}, 
-{0x02150001, 0x01820502, "Animal Crossing", "Animal Crossing", "Pinky", "Pinky", ""}, 
-{0x02160001, 0x00570502, "Animal Crossing", "Animal Crossing", "Curt", "Curt", ""}, 
-{0x02170001, 0x01b30502, "Animal Crossing", "Animal Crossing", "Chow", "Chow", ""}, 
-{0x02190001, 0x007e0502, "Animal Crossing", "Animal Crossing", "Nate", "Nate", ""}, 
-{0x021a0001, 0x00da0502, "Animal Crossing", "Animal Crossing", "Groucho", "Groucho", ""}, 
-{0x021b0001, 0x00800502, "Animal Crossing", "Animal Crossing", "Tutu", "Tutu", ""}, 
-{0x021c0001, 0x02f70502, "Animal Crossing", "Animal Crossing", "Ursala", "Ursala", ""}, 
-{0x021d0001, 0x01cd0502, "Animal Crossing", "Animal Crossing", "Grizzly", "Grizzly", ""}, 
-{0x021e0001, 0x01230502, "Animal Crossing", "Animal Crossing", "Paula", "Paula", ""}, 
-{0x021f0001, 0x03170502, "Animal Crossing", "Animal Crossing", "Ike", "Ike", ""}, 
-{0x02200001, 0x00fd0502, "Animal Crossing", "Animal Crossing", "Charlise", "Charlise", ""}, 
-{0x02210001, 0x013c0502, "Animal Crossing", "Animal Crossing", "Beardo", "Beardo", ""}, 
-{0x02220001, 0x01440502, "Animal Crossing", "Animal Crossing", "Klaus", "Klaus", ""}, 
-{0x022d0001, 0x00f20502, "Animal Crossing", "Animal Crossing", "Jay", "Jay", ""}, 
-{0x022e0001, 0x01d30502, "Animal Crossing", "Animal Crossing", "Robin", "Robin", ""}, 
-{0x022f0001, 0x011e0502, "Animal Crossing", "Animal Crossing", "Anchovy", "Anchovy", ""}, 
-{0x02300001, 0x01d20502, "Animal Crossing", "Animal Crossing", "Twiggy", "Twiggy", ""}, 
-{0x02310001, 0x006a0502, "Animal Crossing", "Animal Crossing", "Jitters", "Jitters", ""}, 
-{0x02320001, 0x02ea0502, "Animal Crossing", "Animal Crossing", "Piper", "Piper", ""}, 
-{0x02330001, 0x03060502, "Animal Crossing", "Animal Crossing", "Admiral", "Admiral", ""}, 
-{0x02350001, 0x00840502, "Animal Crossing", "Animal Crossing", "Midge", "Midge", ""}, 
-{0x02380001, 0x02f80502, "Animal Crossing", "Animal Crossing", "Jacob", "Jacob", ""}, 
-{0x023c0001, 0x00bd0502, "Animal Crossing", "Animal Crossing", "Lucha", "Lucha", ""}, 
-{0x023d0001, 0x01b50502, "Animal Crossing", "Animal Crossing", "Jacques", "Jacques", ""}, 
-{0x023e0001, 0x00d10502, "Animal Crossing", "Animal Crossing", "Peck", "Peck", ""}, 
-{0x023f0001, 0x01660502, "Animal Crossing", "Animal Crossing", "Sparro", "Sparro", ""}, 
-{0x024a0001, 0x01d10502, "Animal Crossing", "Animal Crossing", "Angus", "Angus", ""}, 
-{0x024b0001, 0x01260502, "Animal Crossing", "Animal Crossing", "Rodeo", "Rodeo", ""}, 
-{0x024d0001, 0x02f60502, "Animal Crossing", "Animal Crossing", "Stu", "Stu", ""}, 
-{0x024f0001, 0x00810502, "Animal Crossing", "Animal Crossing", "T-Bone", "T-Bone", ""}, 
-{0x02510001, 0x00c10502, "Animal Crossing", "Animal Crossing", "Coach", "Coach", ""}, 
-{0x02520001, 0x00fe0502, "Animal Crossing", "Animal Crossing", "Vic", "Vic", ""}, 
-{0x025d0001, 0x00550502, "Animal Crossing", "Animal Crossing", "Bob", "Bob", ""}, 
-{0x025e0001, 0x01250502, "Animal Crossing", "Animal Crossing", "Mitzi", "Mitzi", ""}, 
-{0x025f0001, 0x01c50502, "Animal Crossing", "Animal Crossing", "Rosie", "Rosie", ""}, 
-{0x025f0001, 0x01d70502, "Animal Crossing", "Animal Crossing", "Rosie", "Rosie - Amiibo Festival", ""}, 
-{0x02600001, 0x00d20502, "Animal Crossing", "Animal Crossing", "Olivia", "Olivia", ""}, 
-{0x02610001, 0x00650502, "Animal Crossing", "Animal Crossing", "Kiki", "Kiki", ""}, 
-{0x02620001, 0x01370502, "Animal Crossing", "Animal Crossing", "Tangy", "Tangy", ""}, 
-{0x02630001, 0x00750502, "Animal Crossing", "Animal Crossing", "Punchy", "Punchy", ""}, 
-{0x02640001, 0x01ac0502, "Animal Crossing", "Animal Crossing", "Purrl", "Purrl", ""}, 
-{0x02650001, 0x01540502, "Animal Crossing", "Animal Crossing", "Moe", "Moe", ""}, 
-{0x02660001, 0x00680502, "Animal Crossing", "Animal Crossing", "Kabuki", "Kabuki", ""}, 
-{0x02670001, 0x01080502, "Animal Crossing", "Animal Crossing", "Kid Cat", "Kid Cat", ""}, 
-{0x02680001, 0x007d0502, "Animal Crossing", "Animal Crossing", "Monique", "Monique", ""}, 
-{0x02690001, 0x011f0502, "Animal Crossing", "Animal Crossing", "Tabby", "Tabby", ""}, 
-{0x026a0001, 0x01460502, "Animal Crossing", "Animal Crossing", "Stinky", "Stinky", ""}, 
-{0x026b0001, 0x00e90502, "Animal Crossing", "Animal Crossing", "Kitty", "Kitty", ""}, 
-{0x026c0001, 0x00c30502, "Animal Crossing", "Animal Crossing", "Tom", "Tom", ""}, 
-{0x026d0001, 0x013f0502, "Animal Crossing", "Animal Crossing", "Merry", "Merry", ""}, 
-{0x026e0001, 0x00ba0502, "Animal Crossing", "Animal Crossing", "Felicity", "Felicity", ""}, 
-{0x026f0001, 0x01900502, "Animal Crossing", "Animal Crossing", "Lolly", "Lolly", ""}, 
-{0x02700001, 0x00ff0502, "Animal Crossing", "Animal Crossing", "Ankha", "Ankha", ""}, 
-{0x02710001, 0x019b0502, "Animal Crossing", "Animal Crossing", "Rudy", "Rudy", ""}, 
-{0x02720001, 0x01860502, "Animal Crossing", "Animal Crossing", "Katt", "Katt", ""}, 
-{0x027d0001, 0x00630502, "Animal Crossing", "Animal Crossing", "Bluebear", "Bluebear", ""}, 
-{0x027e0001, 0x01690502, "Animal Crossing", "Animal Crossing", "Maple", "Maple", ""}, 
-{0x027f0001, 0x00b90502, "Animal Crossing", "Animal Crossing", "Poncho", "Poncho", ""}, 
-{0x02800001, 0x00830502, "Animal Crossing", "Animal Crossing", "Pudge", "Pudge", ""}, 
-{0x02810001, 0x01200502, "Animal Crossing", "Animal Crossing", "Kody", "Kody", ""}, 
-{0x02820001, 0x01810502, "Animal Crossing", "Animal Crossing", "Stitches", "Stitches", ""}, 
-{0x02820001, 0x01d60502, "Animal Crossing", "Animal Crossing", "Stitches", "Stitches - Amiibo Festival", ""}, 
-{0x02830001, 0x00c70502, "Animal Crossing", "Animal Crossing", "Vladimir", "Vladimir", ""}, 
-{0x02840001, 0x02fe0502, "Animal Crossing", "Animal Crossing", "Murphy", "Murphy", ""}, 
-{0x02860001, 0x03130502, "Animal Crossing", "Animal Crossing", "Olive", "Olive", ""}, 
-{0x02870001, 0x005a0502, "Animal Crossing", "Animal Crossing", "Cheri", "Cheri", ""}, 
-{0x028a0001, 0x02e90502, "Animal Crossing", "Animal Crossing", "June", "June", ""}, 
-{0x028b0001, 0x00e30502, "Animal Crossing", "Animal Crossing", "Pekoe", "Pekoe", ""}, 
-{0x028c0001, 0x013e0502, "Animal Crossing", "Animal Crossing", "Chester", "Chester", ""}, 
-{0x028d0001, 0x01bd0502, "Animal Crossing", "Animal Crossing", "Barold", "Barold", ""}, 
-{0x028e0001, 0x019e0502, "Animal Crossing", "Animal Crossing", "Tammy", "Tammy", ""}, 
-{0x028f0101, 0x031a0502, "Animal Crossing", "Animal Crossing", "Marty", "Marty", ""}, 
-{0x02990001, 0x00950502, "Animal Crossing", "Animal Crossing", "Goose", "Goose", ""}, 
-{0x029a0001, 0x00ee0502, "Animal Crossing", "Animal Crossing", "Benedict", "Benedict", ""}, 
-{0x029b0001, 0x00cb0502, "Animal Crossing", "Animal Crossing", "Egbert", "Egbert", ""}, 
-{0x029e0001, 0x013d0502, "Animal Crossing", "Animal Crossing", "Ava", "Ava", ""}, 
-{0x02a20001, 0x01ba0502, "Animal Crossing", "Animal Crossing", "Becky", "Becky", ""}, 
-{0x02a30001, 0x02ff0502, "Animal Crossing", "Animal Crossing", "Plucky", "Plucky", ""}, 
-{0x02a40001, 0x00720502, "Animal Crossing", "Animal Crossing", "Knox", "Knox", ""}, 
-{0x02a50001, 0x018c0502, "Animal Crossing", "Animal Crossing", "Broffina", "Broffina", ""}, 
-{0x02a60001, 0x01240502, "Animal Crossing", "Animal Crossing", "Ken", "Ken", ""}, 
-{0x02b10001, 0x00690502, "Animal Crossing", "Animal Crossing", "Patty", "Patty", ""}, 
-{0x02b20001, 0x00c40502, "Animal Crossing", "Animal Crossing", "Tipper", "Tipper", ""}, 
-{0x02b70001, 0x030f0502, "Animal Crossing", "Animal Crossing", "Norma", "Norma", ""}, 
-{0x02b80001, 0x019c0502, "Animal Crossing", "Animal Crossing", "Naomi", "Naomi", ""}, 
-{0x02c30001, 0x00dc0502, "Animal Crossing", "Animal Crossing", "Alfonso", "Alfonso", ""}, 
-{0x02c40001, 0x00670502, "Animal Crossing", "Animal Crossing", "Alli", "Alli", ""}, 
-{0x02c50001, 0x03080502, "Animal Crossing", "Animal Crossing", "Boots", "Boots", ""}, 
-{0x02c70001, 0x01220502, "Animal Crossing", "Animal Crossing", "Del", "Del", ""}, 
-{0x02c90001, 0x00cd0502, "Animal Crossing", "Animal Crossing", "Sly", "Sly", ""}, 
-{0x02ca0001, 0x01ca0502, "Animal Crossing", "Animal Crossing", "Gayle", "Gayle", ""}, 
-{0x02cb0001, 0x01360502, "Animal Crossing", "Animal Crossing", "Drago", "Drago", ""}, 
-{0x02d60001, 0x00560502, "Animal Crossing", "Animal Crossing", "Fauna", "Fauna", ""}, 
-{0x02d70001, 0x01300502, "Animal Crossing", "Animal Crossing", "Bam", "Bam", ""}, 
-{0x02d80001, 0x00e20502, "Animal Crossing", "Animal Crossing", "Zell", "Zell", ""}, 
-{0x02d90001, 0x01c80502, "Animal Crossing", "Animal Crossing", "Bruce", "Bruce", ""}, 
-{0x02da0001, 0x01330502, "Animal Crossing", "Animal Crossing", "Deirdre", "Deirdre", ""}, 
-{0x02db0001, 0x005e0502, "Animal Crossing", "Animal Crossing", "Lopez", "Lopez", ""}, 
-{0x02dc0001, 0x00be0502, "Animal Crossing", "Animal Crossing", "Fuchsia", "Fuchsia", ""}, 
-{0x02dd0001, 0x00ea0502, "Animal Crossing", "Animal Crossing", "Beau", "Beau", ""}, 
-{0x02de0001, 0x009c0502, "Animal Crossing", "Animal Crossing", "Diana", "Diana", ""}, 
-{0x02df0001, 0x01910502, "Animal Crossing", "Animal Crossing", "Erik", "Erik", ""}, 
-{0x02e00101, 0x031d0502, "Animal Crossing", "Animal Crossing", "Chelsea", "Chelsea", ""}, 
-{0x02ea0001, 0x01800502, "Animal Crossing", "Animal Crossing", "Goldie", "Goldie", ""}, 
-{0x02ea0001, 0x01d50502, "Animal Crossing", "Animal Crossing", "Goldie", "Goldie - Amiibo Festival", ""}, 
-{0x02eb0001, 0x00de0502, "Animal Crossing", "Animal Crossing", "Butch", "Butch", ""}, 
-{0x02ec0001, 0x01c40502, "Animal Crossing", "Animal Crossing", "Lucky", "Lucky", ""}, 
-{0x02ed0001, 0x015a0502, "Animal Crossing", "Animal Crossing", "Biskit", "Biskit", ""}, 
-{0x02ee0001, 0x01990502, "Animal Crossing", "Animal Crossing", "Bones", "Bones", ""}, 
-{0x02ef0001, 0x00580502, "Animal Crossing", "Animal Crossing", "Portia", "Portia", ""}, 
-{0x02f00001, 0x00a70502, "Animal Crossing", "Animal Crossing", "Walker", "Walker", ""}, 
-{0x02f10001, 0x01450502, "Animal Crossing", "Animal Crossing", "Daisy", "Daisy", ""}, 
-{0x02f20001, 0x00cc0502, "Animal Crossing", "Animal Crossing", "Cookie", "Cookie", ""}, 
-{0x02f30001, 0x02f90502, "Animal Crossing", "Animal Crossing", "Maddie", "Maddie", ""}, 
-{0x02f40001, 0x03050502, "Animal Crossing", "Animal Crossing", "Bea", "Bea", ""}, 
-{0x02f80001, 0x01380502, "Animal Crossing", "Animal Crossing", "Mac", "Mac", ""}, 
-{0x02f90001, 0x01020502, "Animal Crossing", "Animal Crossing", "Marcel", "Marcel", ""}, 
-{0x02fa0001, 0x00970502, "Animal Crossing", "Animal Crossing", "Benjamin", "Benjamin", ""}, 
-{0x02fb0001, 0x00900502, "Animal Crossing", "Animal Crossing", "Cherry", "Cherry", ""}, 
-{0x02fc0001, 0x018f0502, "Animal Crossing", "Animal Crossing", "Shep", "Shep", ""}, 
-{0x03070001, 0x00640502, "Animal Crossing", "Animal Crossing", "Bill", "Bill", ""}, 
-{0x03080001, 0x014d0502, "Animal Crossing", "Animal Crossing", "Joey", "Joey", ""}, 
-{0x03090001, 0x00c60502, "Animal Crossing", "Animal Crossing", "Pate", "Pate", ""}, 
-{0x030a0001, 0x01c70502, "Animal Crossing", "Animal Crossing", "Maelle", "Maelle", ""}, 
-{0x030b0001, 0x00790502, "Animal Crossing", "Animal Crossing", "Deena", "Deena", ""}, 
-{0x030c0001, 0x01b80502, "Animal Crossing", "Animal Crossing", "Pompom", "Pompom", ""}, 
-{0x030d0001, 0x01840502, "Animal Crossing", "Animal Crossing", "Mallary", "Mallary", ""}, 
-{0x030e0001, 0x012f0502, "Animal Crossing", "Animal Crossing", "Freckles", "Freckles", ""}, 
-{0x030f0001, 0x016d0502, "Animal Crossing", "Animal Crossing", "Derwin", "Derwin", ""}, 
-{0x03100001, 0x00f80502, "Animal Crossing", "Animal Crossing", "Drake", "Drake", ""}, 
-{0x03110001, 0x00d60502, "Animal Crossing", "Animal Crossing", "Scoot", "Scoot", ""}, 
-{0x03120001, 0x03090502, "Animal Crossing", "Animal Crossing", "Weber", "Weber", ""}, 
-{0x03130001, 0x01210502, "Animal Crossing", "Animal Crossing", "Miranda", "Miranda", ""}, 
-{0x03140001, 0x02f40502, "Animal Crossing", "Animal Crossing", "Ketchup", "Ketchup", ""}, 
-{0x03160001, 0x01c00502, "Animal Crossing", "Animal Crossing", "Gloria", "Gloria", ""}, 
-{0x03170001, 0x00a60502, "Animal Crossing", "Animal Crossing", "Molly", "Molly", ""}, 
-{0x03180001, 0x006c0502, "Animal Crossing", "Animal Crossing", "Quillson", "Quillson", ""}, 
-{0x03230001, 0x00760502, "Animal Crossing", "Animal Crossing", "Opal", "Opal", ""}, 
-{0x03240001, 0x01890502, "Animal Crossing", "Animal Crossing", "Dizzy", "Dizzy", ""}, 
-{0x03250001, 0x010a0502, "Animal Crossing", "Animal Crossing", "Big Top", "Big Top", ""}, 
-{0x03260001, 0x01390502, "Animal Crossing", "Animal Crossing", "Eloise", "Eloise", ""}, 
-{0x03270001, 0x01c30502, "Animal Crossing", "Animal Crossing", "Margie", "Margie", ""}, 
-{0x03280001, 0x02eb0502, "Animal Crossing", "Animal Crossing", "Paolo", "Paolo", ""}, 
-{0x03290001, 0x009d0502, "Animal Crossing", "Animal Crossing", "Axel", "Axel", ""}, 
-{0x032a0001, 0x03070502, "Animal Crossing", "Animal Crossing", "Ellie", "Ellie", ""}, 
-{0x032c0001, 0x01480502, "Animal Crossing", "Animal Crossing", "Tucker", "Tucker", ""}, 
-{0x032d0001, 0x00bc0502, "Animal Crossing", "Animal Crossing", "Tia", "Tia", ""}, 
-{0x032e0101, 0x031c0502, "Animal Crossing", "Animal Crossing", "Chai", "Chai", ""}, 
-{0x03380001, 0x011d0502, "Animal Crossing", "Animal Crossing", "Lily", "Lily", ""}, 
-{0x03390001, 0x01b10502, "Animal Crossing", "Animal Crossing", "Ribbot", "Ribbot", ""}, 
-{0x033a0001, 0x01cc0502, "Animal Crossing", "Animal Crossing", "Frobert", "Frobert", ""}, 
-{0x033b0001, 0x00fa0502, "Animal Crossing", "Animal Crossing", "Camofrog", "Camofrog", ""}, 
-{0x033c0001, 0x01000502, "Animal Crossing", "Animal Crossing", "Drift", "Drift", ""}, 
-{0x033d0001, 0x013a0502, "Animal Crossing", "Animal Crossing", "Wart Jr.", "Wart Jr.", ""}, 
-{0x033e0001, 0x01a20502, "Animal Crossing", "Animal Crossing", "Puddles", "Puddles", ""}, 
-{0x033f0001, 0x008f0502, "Animal Crossing", "Animal Crossing", "Jeremiah", "Jeremiah", ""}, 
-{0x03410001, 0x030e0502, "Animal Crossing", "Animal Crossing", "Tad", "Tad", ""}, 
-{0x03420001, 0x01280502, "Animal Crossing", "Animal Crossing", "Cousteau", "Cousteau", ""}, 
-{0x03430001, 0x02ef0502, "Animal Crossing", "Animal Crossing", "Huck", "Huck", ""}, 
-{0x03440001, 0x00c50502, "Animal Crossing", "Animal Crossing", "Prince", "Prince", ""}, 
-{0x03450001, 0x005f0502, "Animal Crossing", "Animal Crossing", "Jambette", "Jambette", ""}, 
-{0x03470001, 0x03020502, "Animal Crossing", "Animal Crossing", "Raddle", "Raddle", ""}, 
-{0x03480001, 0x006b0502, "Animal Crossing", "Animal Crossing", "Gigi", "Gigi", ""}, 
-{0x03490001, 0x018d0502, "Animal Crossing", "Animal Crossing", "Croque", "Croque", ""}, 
-{0x034a0001, 0x01430502, "Animal Crossing", "Animal Crossing", "Diva", "Diva", ""}, 
-{0x034b0001, 0x009f0502, "Animal Crossing", "Animal Crossing", "Henry", "Henry", ""}, 
-{0x03560001, 0x01350502, "Animal Crossing", "Animal Crossing", "Chevre", "Chevre", ""}, 
-{0x03570001, 0x00eb0502, "Animal Crossing", "Animal Crossing", "Nan", "Nan", ""}, 
-{0x03580001, 0x02fa0502, "Animal Crossing", "Animal Crossing", "Billy", "Billy", ""}, 
-{0x035a0001, 0x00850502, "Animal Crossing", "Animal Crossing", "Gruff", "Gruff", ""}, 
-{0x035c0001, 0x01290502, "Animal Crossing", "Animal Crossing", "Velma", "Velma", ""}, 
-{0x035d0001, 0x00c90502, "Animal Crossing", "Animal Crossing", "Kidd", "Kidd", ""}, 
-{0x035e0001, 0x018e0502, "Animal Crossing", "Animal Crossing", "Pashmina", "Pashmina", ""}, 
-{0x03690001, 0x00d30502, "Animal Crossing", "Animal Crossing", "Cesar", "Cesar", ""}, 
-{0x036a0001, 0x019d0502, "Animal Crossing", "Animal Crossing", "Peewee", "Peewee", ""}, 
-{0x036b0001, 0x018b0502, "Animal Crossing", "Animal Crossing", "Boone", "Boone", ""}, 
-{0x036d0001, 0x03040502, "Animal Crossing", "Animal Crossing", "Louie", "Louie", ""}, 
-{0x036e0001, 0x02fb0502, "Animal Crossing", "Animal Crossing", "Boyd", "Boyd", ""}, 
-{0x03700001, 0x015d0502, "Animal Crossing", "Animal Crossing", "Violet", "Violet", ""}, 
-{0x03710001, 0x005c0502, "Animal Crossing", "Animal Crossing", "Al", "Al", ""}, 
-{0x03720001, 0x010b0502, "Animal Crossing", "Animal Crossing", "Rocket", "Rocket", ""}, 
-{0x03730001, 0x01340502, "Animal Crossing", "Animal Crossing", "Hans", "Hans", ""}, 
-{0x03740101, 0x03190502, "Animal Crossing", "Animal Crossing", "Rilla", "Rilla", ""}, 
-{0x037e0001, 0x01560502, "Animal Crossing", "Animal Crossing", "Hamlet", "Hamlet", ""}, 
-{0x037f0001, 0x01aa0502, "Animal Crossing", "Animal Crossing", "Apple", "Apple", ""}, 
-{0x03800001, 0x01870502, "Animal Crossing", "Animal Crossing", "Graham", "Graham", ""}, 
-{0x03810001, 0x00d50502, "Animal Crossing", "Animal Crossing", "Rodney", "Rodney", ""}, 
-{0x03820001, 0x016b0502, "Animal Crossing", "Animal Crossing", "Soleil", "Soleil", ""}, 
-{0x03830001, 0x009b0502, "Animal Crossing", "Animal Crossing", "Clay", "Clay", ""}, 
-{0x03840001, 0x00860502, "Animal Crossing", "Animal Crossing", "Flurry", "Flurry", ""}, 
-{0x03850001, 0x01060502, "Animal Crossing", "Animal Crossing", "Hamphrey", "Hamphrey", ""}, 
-{0x03900001, 0x01850502, "Animal Crossing", "Animal Crossing", "Rocco", "Rocco", ""}, 
-{0x03920001, 0x01270502, "Animal Crossing", "Animal Crossing", "Bubbles", "Bubbles", ""}, 
-{0x03930001, 0x00a00502, "Animal Crossing", "Animal Crossing", "Bertha", "Bertha", ""}, 
-{0x03940001, 0x00890502, "Animal Crossing", "Animal Crossing", "Biff", "Biff", ""}, 
-{0x03950001, 0x02fc0502, "Animal Crossing", "Animal Crossing", "Bitty", "Bitty", ""}, 
-{0x03980001, 0x00bf0502, "Animal Crossing", "Animal Crossing", "Harry", "Harry", ""}, 
-{0x03990001, 0x01c20502, "Animal Crossing", "Animal Crossing", "Hippeux", "Hippeux", ""}, 
-{0x03a40001, 0x014f0502, "Animal Crossing", "Animal Crossing", "Buck", "Buck", ""}, 
-{0x03a50001, 0x015b0502, "Animal Crossing", "Animal Crossing", "Victoria", "Victoria", ""}, 
-{0x03a60001, 0x00c80502, "Animal Crossing", "Animal Crossing", "Savannah", "Savannah", ""}, 
-{0x03a70001, 0x01a10502, "Animal Crossing", "Animal Crossing", "Elmer", "Elmer", ""}, 
-{0x03a80001, 0x00910502, "Animal Crossing", "Animal Crossing", "Rosco", "Rosco", ""}, 
-{0x03a90001, 0x00710502, "Animal Crossing", "Animal Crossing", "Winnie", "Winnie", ""}, 
-{0x03aa0001, 0x00e60502, "Animal Crossing", "Animal Crossing", "Ed", "Ed", ""}, 
-{0x03ab0001, 0x03160502, "Animal Crossing", "Animal Crossing", "Cleo", "Cleo", ""}, 
-{0x03ac0001, 0x01880502, "Animal Crossing", "Animal Crossing", "Peaches", "Peaches", ""}, 
-{0x03ad0001, 0x01b20502, "Animal Crossing", "Animal Crossing", "Annalise", "Annalise", ""}, 
-{0x03ae0001, 0x00870502, "Animal Crossing", "Animal Crossing", "Clyde", "Clyde", ""}, 
-{0x03af0001, 0x012c0502, "Animal Crossing", "Animal Crossing", "Colton", "Colton", ""}, 
-{0x03b00001, 0x01a90502, "Animal Crossing", "Animal Crossing", "Papi", "Papi", ""}, 
-{0x03b10001, 0x00f00502, "Animal Crossing", "Animal Crossing", "Julian", "Julian", ""}, 
-{0x03bc0001, 0x008a0502, "Animal Crossing", "Animal Crossing", "Yuka", "Yuka", ""}, 
-{0x03bd0001, 0x00f90502, "Animal Crossing", "Animal Crossing", "Alice", "Alice", ""}, 
-{0x03be0001, 0x01980502, "Animal Crossing", "Animal Crossing", "Melba", "Melba", ""}, 
-{0x03bf0001, 0x01bc0502, "Animal Crossing", "Animal Crossing", "Sydney", "Sydney", ""}, 
-{0x03c00001, 0x03100502, "Animal Crossing", "Animal Crossing", "Gonzo", "Gonzo", ""}, 
-{0x03c10001, 0x00bb0502, "Animal Crossing", "Animal Crossing", "Ozzie", "Ozzie", ""}, 
-{0x03c40001, 0x012b0502, "Animal Crossing", "Animal Crossing", "Canberra", "Canberra", ""}, 
-{0x03c50001, 0x015c0502, "Animal Crossing", "Animal Crossing", "Lyman", "Lyman", ""}, 
-{0x03c60001, 0x00930502, "Animal Crossing", "Animal Crossing", "Eugene", "Eugene", ""}, 
-{0x03d10001, 0x00c20502, "Animal Crossing", "Animal Crossing", "Kitt", "Kitt", ""}, 
-{0x03d20001, 0x00e50502, "Animal Crossing", "Animal Crossing", "Mathilda", "Mathilda", ""}, 
-{0x03d30001, 0x02f30502, "Animal Crossing", "Animal Crossing", "Carrie", "Carrie", ""}, 
-{0x03d60001, 0x01570502, "Animal Crossing", "Animal Crossing", "Astrid", "Astrid", ""}, 
-{0x03d70001, 0x01b40502, "Animal Crossing", "Animal Crossing", "Sylvia", "Sylvia", ""}, 
-{0x03d90001, 0x01a50502, "Animal Crossing", "Animal Crossing", "Walt", "Walt", ""}, 
-{0x03da0001, 0x01510502, "Animal Crossing", "Animal Crossing", "Rooney", "Rooney", ""}, 
-{0x03db0001, 0x006d0502, "Animal Crossing", "Animal Crossing", "Marcie", "Marcie", ""}, 
-{0x03e60001, 0x00ec0502, "Animal Crossing", "Animal Crossing", "Bud", "Bud", ""}, 
-{0x03e70001, 0x012a0502, "Animal Crossing", "Animal Crossing", "Elvis", "Elvis", ""}, 
-{0x03e80001, 0x02f50502, "Animal Crossing", "Animal Crossing", "Rex", "Rex", ""}, 
-{0x03ea0001, 0x030b0502, "Animal Crossing", "Animal Crossing", "Leopold", "Leopold", ""}, 
-{0x03ec0001, 0x01830502, "Animal Crossing", "Animal Crossing", "Mott", "Mott", ""}, 
-{0x03ed0001, 0x01a30502, "Animal Crossing", "Animal Crossing", "Rory", "Rory", ""}, 
-{0x03ee0001, 0x008b0502, "Animal Crossing", "Animal Crossing", "Lionel", "Lionel", ""}, 
-{0x03fa0001, 0x00d00502, "Animal Crossing", "Animal Crossing", "Nana", "Nana", ""}, 
-{0x03fb0001, 0x01cf0502, "Animal Crossing", "Animal Crossing", "Simon", "Simon", ""}, 
-{0x03fc0001, 0x01470502, "Animal Crossing", "Animal Crossing", "Tammi", "Tammi", ""}, 
-{0x03fd0001, 0x01580502, "Animal Crossing", "Animal Crossing", "Monty", "Monty", ""}, 
-{0x03fe0001, 0x01a40502, "Animal Crossing", "Animal Crossing", "Elise", "Elise", ""}, 
-{0x03ff0001, 0x00f40502, "Animal Crossing", "Animal Crossing", "Flip", "Flip", ""}, 
-{0x04000001, 0x006f0502, "Animal Crossing", "Animal Crossing", "Shari", "Shari", ""}, 
-{0x04010001, 0x00660502, "Animal Crossing", "Animal Crossing", "Deli", "Deli", ""}, 
-{0x040c0001, 0x01590502, "Animal Crossing", "Animal Crossing", "Dora", "Dora", ""}, 
-{0x040d0001, 0x00780502, "Animal Crossing", "Animal Crossing", "Limberg", "Limberg", ""}, 
-{0x040e0001, 0x00880502, "Animal Crossing", "Animal Crossing", "Bella", "Bella", ""}, 
-{0x040f0001, 0x01500502, "Animal Crossing", "Animal Crossing", "Bree", "Bree", ""}, 
-{0x04100001, 0x007f0502, "Animal Crossing", "Animal Crossing", "Samson", "Samson", ""}, 
-{0x04110001, 0x01ab0502, "Animal Crossing", "Animal Crossing", "Rod", "Rod", ""}, 
-{0x04140001, 0x030a0502, "Animal Crossing", "Animal Crossing", "Candi", "Candi", ""}, 
-{0x04150001, 0x01bb0502, "Animal Crossing", "Animal Crossing", "Rizzo", "Rizzo", ""}, 
-{0x04160001, 0x00fb0502, "Animal Crossing", "Animal Crossing", "Anicotti", "Anicotti", ""}, 
-{0x04180001, 0x00d80502, "Animal Crossing", "Animal Crossing", "Broccolo", "Broccolo", ""}, 
-{0x041a0001, 0x00e00502, "Animal Crossing", "Animal Crossing", "Moose", "Moose", ""}, 
-{0x041b0001, 0x00f10502, "Animal Crossing", "Animal Crossing", "Bettina", "Bettina", ""}, 
-{0x041c0001, 0x01410502, "Animal Crossing", "Animal Crossing", "Greta", "Greta", ""}, 
-{0x041d0001, 0x018a0502, "Animal Crossing", "Animal Crossing", "Penelope", "Penelope", ""}, 
-{0x041e0001, 0x015f0502, "Animal Crossing", "Animal Crossing", "Chadder", "Chadder", ""}, 
-{0x04290001, 0x00700502, "Animal Crossing", "Animal Crossing", "Octavian", "Octavian", ""}, 
-{0x042a0001, 0x012d0502, "Animal Crossing", "Animal Crossing", "Marina", "Marina", ""}, 
-{0x042b0001, 0x01af0502, "Animal Crossing", "Animal Crossing", "Zucker", "Zucker", ""}, 
-{0x04360001, 0x01940502, "Animal Crossing", "Animal Crossing", "Queenie", "Queenie", ""}, 
-{0x04370001, 0x01050502, "Animal Crossing", "Animal Crossing", "Gladys", "Gladys", ""}, 
-{0x04380001, 0x03000502, "Animal Crossing", "Animal Crossing", "Sandy", "Sandy", ""}, 
-{0x04390001, 0x03110502, "Animal Crossing", "Animal Crossing", "Sprocket", "Sprocket", ""}, 
-{0x043b0001, 0x03030502, "Animal Crossing", "Animal Crossing", "Julia", "Julia", ""}, 
-{0x043c0001, 0x01cb0502, "Animal Crossing", "Animal Crossing", "Cranston", "Cranston", ""}, 
-{0x043d0001, 0x007c0502, "Animal Crossing", "Animal Crossing", "Phil", "Phil", ""}, 
-{0x043e0001, 0x01490502, "Animal Crossing", "Animal Crossing", "Blanche", "Blanche", ""}, 
-{0x043f0001, 0x01550502, "Animal Crossing", "Animal Crossing", "Flora", "Flora", ""}, 
-{0x04400001, 0x00ca0502, "Animal Crossing", "Animal Crossing", "Phoebe", "Phoebe", ""}, 
-{0x044b0001, 0x016c0502, "Animal Crossing", "Animal Crossing", "Apollo", "Apollo", ""}, 
-{0x044c0001, 0x008e0502, "Animal Crossing", "Animal Crossing", "Amelia", "Amelia", ""}, 
-{0x044d0001, 0x01930502, "Animal Crossing", "Animal Crossing", "Pierce", "Pierce", ""}, 
-{0x044e0001, 0x03150502, "Animal Crossing", "Animal Crossing", "Buzz", "Buzz", ""}, 
-{0x04500001, 0x00cf0502, "Animal Crossing", "Animal Crossing", "Avery", "Avery", ""}, 
-{0x04510001, 0x015e0502, "Animal Crossing", "Animal Crossing", "Frank", "Frank", ""}, 
-{0x04520001, 0x00730502, "Animal Crossing", "Animal Crossing", "Sterling", "Sterling", ""}, 
-{0x04530001, 0x01040502, "Animal Crossing", "Animal Crossing", "Keaton", "Keaton", ""}, 
-{0x04540001, 0x01ae0502, "Animal Crossing", "Animal Crossing", "Celia", "Celia", ""}, 
-{0x045f0001, 0x01a80502, "Animal Crossing", "Animal Crossing", "Aurora", "Aurora", ""}, 
-{0x04600001, 0x00a50502, "Animal Crossing", "Animal Crossing", "Roald", "Roald", ""}, 
-{0x04610001, 0x01610502, "Animal Crossing", "Animal Crossing", "Cube", "Cube", ""}, 
-{0x04620001, 0x00f60502, "Animal Crossing", "Animal Crossing", "Hopper", "Hopper", ""}, 
-{0x04630001, 0x01310502, "Animal Crossing", "Animal Crossing", "Friga", "Friga", ""}, 
-{0x04640001, 0x00c00502, "Animal Crossing", "Animal Crossing", "Gwen", "Gwen", ""}, 
-{0x04650001, 0x006e0502, "Animal Crossing", "Animal Crossing", "Puck", "Puck", ""}, 
-{0x04680001, 0x02f20502, "Animal Crossing", "Animal Crossing", "Wade", "Wade", ""}, 
-{0x04690001, 0x01640502, "Animal Crossing", "Animal Crossing", "Boomer", "Boomer", ""}, 
-{0x046a0001, 0x01d00502, "Animal Crossing", "Animal Crossing", "Iggly", "Iggly", ""}, 
-{0x046b0001, 0x01970502, "Animal Crossing", "Animal Crossing", "Tex", "Tex", ""}, 
-{0x046c0001, 0x008c0502, "Animal Crossing", "Animal Crossing", "Flo", "Flo", ""}, 
-{0x046d0001, 0x00f30502, "Animal Crossing", "Animal Crossing", "Sprinkle", "Sprinkle", ""}, 
-{0x04780001, 0x01630502, "Animal Crossing", "Animal Crossing", "Curly", "Curly", ""}, 
-{0x04790001, 0x00920502, "Animal Crossing", "Animal Crossing", "Truffles", "Truffles", ""}, 
-{0x047a0001, 0x00600502, "Animal Crossing", "Animal Crossing", "Rasher", "Rasher", ""}, 
-{0x047b0001, 0x00f50502, "Animal Crossing", "Animal Crossing", "Hugh", "Hugh", ""}, 
-{0x047c0001, 0x01a00502, "Animal Crossing", "Animal Crossing", "Lucy", "Lucy", ""}, 
-{0x047d0001, 0x012e0502, "Animal Crossing", "Animal Crossing", "Spork/Crackle", "Spork/Crackle", ""}, 
-{0x04800001, 0x008d0502, "Animal Crossing", "Animal Crossing", "Cobb", "Cobb", ""}, 
-{0x04810001, 0x02f10502, "Animal Crossing", "Animal Crossing", "Boris", "Boris", ""}, 
-{0x04820001, 0x02fd0502, "Animal Crossing", "Animal Crossing", "Maggie", "Maggie", ""}, 
-{0x04830001, 0x01b00502, "Animal Crossing", "Animal Crossing", "Peggy", "Peggy", ""}, 
-{0x04850001, 0x014c0502, "Animal Crossing", "Animal Crossing", "Gala", "Gala", ""}, 
-{0x04860001, 0x00fc0502, "Animal Crossing", "Animal Crossing", "Chops", "Chops", ""}, 
-{0x04870001, 0x01bf0502, "Animal Crossing", "Animal Crossing", "Kevin", "Kevin", ""}, 
-{0x04880001, 0x00980502, "Animal Crossing", "Animal Crossing", "Pancetti", "Pancetti", ""}, 
-{0x04890001, 0x00ef0502, "Animal Crossing", "Animal Crossing", "Agnes", "Agnes", ""}, 
-{0x04940001, 0x009a0502, "Animal Crossing", "Animal Crossing", "Bunnie", "Bunnie", ""}, 
-{0x04950001, 0x01920502, "Animal Crossing", "Animal Crossing", "Dotty", "Dotty", ""}, 
-{0x04960001, 0x00d90502, "Animal Crossing", "Animal Crossing", "Coco", "Coco", ""}, 
-{0x04970001, 0x007a0502, "Animal Crossing", "Animal Crossing", "Snake", "Snake", ""}, 
-{0x04980001, 0x014a0502, "Animal Crossing", "Animal Crossing", "Gaston", "Gaston", ""}, 
-{0x04990001, 0x00df0502, "Animal Crossing", "Animal Crossing", "Gabi", "Gabi", ""}, 
-{0x049a0001, 0x014e0502, "Animal Crossing", "Animal Crossing", "Pippy", "Pippy", ""}, 
-{0x049b0001, 0x00610502, "Animal Crossing", "Animal Crossing", "Tiffany", "Tiffany", ""}, 
-{0x049c0001, 0x01400502, "Animal Crossing", "Animal Crossing", "Genji", "Genji", ""}, 
-{0x049d0001, 0x00ed0502, "Animal Crossing", "Animal Crossing", "Ruby", "Ruby", ""}, 
-{0x049e0001, 0x01b70502, "Animal Crossing", "Animal Crossing", "Doc", "Doc", ""}, 
-{0x049f0001, 0x03010502, "Animal Crossing", "Animal Crossing", "Claude", "Claude", ""}, 
-{0x04a00001, 0x016e0502, "Animal Crossing", "Animal Crossing", "Francine", "Francine", ""}, 
-{0x04a10001, 0x016f0502, "Animal Crossing", "Animal Crossing", "Chrissy", "Chrissy", ""}, 
-{0x04a20001, 0x02e80502, "Animal Crossing", "Animal Crossing", "Hopkins", "Hopkins", ""}, 
-{0x04a30001, 0x01c90502, "Animal Crossing", "Animal Crossing", "OHare", "OHare", ""}, 
-{0x04a40001, 0x00d40502, "Animal Crossing", "Animal Crossing", "Carmen", "Carmen", ""}, 
-{0x04a50001, 0x00740502, "Animal Crossing", "Animal Crossing", "Bonbon", "Bonbon", ""}, 
-{0x04a60001, 0x00a30502, "Animal Crossing", "Animal Crossing", "Cole", "Cole", ""}, 
-{0x04a70001, 0x01a60502, "Animal Crossing", "Animal Crossing", "Mira", "Mira", ""}, 
-{0x04a80101, 0x031e0502, "Animal Crossing", "Animal Crossing", "Toby", "Toby", ""}, 
-{0x04b20001, 0x01b90502, "Animal Crossing", "Animal Crossing", "Tank", "Tank", ""}, 
-{0x04b30001, 0x00dd0502, "Animal Crossing", "Animal Crossing", "Rhonda", "Rhonda", ""}, 
-{0x04b40001, 0x030c0502, "Animal Crossing", "Animal Crossing", "Spike", "Spike", ""}, 
-{0x04b60001, 0x02ec0502, "Animal Crossing", "Animal Crossing", "Hornsby", "Hornsby", ""}, 
-{0x04b90001, 0x01600502, "Animal Crossing", "Animal Crossing", "Merengue", "Merengue", ""}, 
-{0x04ba0001, 0x005d0502, "Animal Crossing", "Animal Crossing", "Renée", "Renée", ""}, 
-{0x04c50001, 0x01010502, "Animal Crossing", "Animal Crossing", "Vesta", "Vesta", ""}, 
-{0x04c60001, 0x01670502, "Animal Crossing", "Animal Crossing", "Baabara", "Baabara", ""}, 
-{0x04c70001, 0x00940502, "Animal Crossing", "Animal Crossing", "Eunice", "Eunice", ""}, 
-{0x04c80001, 0x02ed0502, "Animal Crossing", "Animal Crossing", "Stella", "Stella", ""}, 
-{0x04c90001, 0x030d0502, "Animal Crossing", "Animal Crossing", "Cashmere", "Cashmere", ""}, 
-{0x04cc0001, 0x00a40502, "Animal Crossing", "Animal Crossing", "Willow", "Willow", ""}, 
-{0x04cd0001, 0x01520502, "Animal Crossing", "Animal Crossing", "Curlos", "Curlos", ""}, 
-{0x04ce0001, 0x00db0502, "Animal Crossing", "Animal Crossing", "Wendy", "Wendy", ""}, 
-{0x04cf0001, 0x00e10502, "Animal Crossing", "Animal Crossing", "Timbra", "Timbra", ""}, 
-{0x04d00001, 0x01960502, "Animal Crossing", "Animal Crossing", "Frita", "Frita", ""}, 
-{0x04d10001, 0x009e0502, "Animal Crossing", "Animal Crossing", "Muffy", "Muffy", ""}, 
-{0x04d20001, 0x01a70502, "Animal Crossing", "Animal Crossing", "Pietro", "Pietro", ""}, 
-{0x04d30101, 0x031b0502, "Animal Crossing", "Animal Crossing", "Étoile", "Étoile", ""}, 
-{0x04dd0001, 0x00a20502, "Animal Crossing", "Animal Crossing", "Peanut", "Peanut", ""}, 
-{0x04de0001, 0x00ce0502, "Animal Crossing", "Animal Crossing", "Blaire", "Blaire", ""}, 
-{0x04df0001, 0x00e80502, "Animal Crossing", "Animal Crossing", "Filbert", "Filbert", ""}, 
-{0x04e00001, 0x00f70502, "Animal Crossing", "Animal Crossing", "Pecan", "Pecan", ""}, 
-{0x04e10001, 0x01be0502, "Animal Crossing", "Animal Crossing", "Nibbles", "Nibbles", ""}, 
-{0x04e20001, 0x01090502, "Animal Crossing", "Animal Crossing", "Agent S", "Agent S", ""}, 
-{0x04e30001, 0x01650502, "Animal Crossing", "Animal Crossing", "Caroline", "Caroline", ""}, 
-{0x04e40001, 0x01b60502, "Animal Crossing", "Animal Crossing", "Sally", "Sally", ""}, 
-{0x04e50001, 0x01ad0502, "Animal Crossing", "Animal Crossing", "Static", "Static", ""}, 
-{0x04e60001, 0x00820502, "Animal Crossing", "Animal Crossing", "Mint", "Mint", ""}, 
-{0x04e70001, 0x01320502, "Animal Crossing", "Animal Crossing", "Ricky", "Ricky", ""}, 
-{0x04e80001, 0x01ce0502, "Animal Crossing", "Animal Crossing", "Cally", "Cally", ""}, 
-{0x04ea0001, 0x03180502, "Animal Crossing", "Animal Crossing", "Tasha", "Tasha", ""}, 
-{0x04eb0001, 0x02f00502, "Animal Crossing", "Animal Crossing", "Sylvana", "Sylvana", ""}, 
-{0x04ec0001, 0x00770502, "Animal Crossing", "Animal Crossing", "Poppy", "Poppy", ""}, 
-{0x04ed0001, 0x00620502, "Animal Crossing", "Animal Crossing", "Sheldon", "Sheldon", ""}, 
-{0x04ee0001, 0x014b0502, "Animal Crossing", "Animal Crossing", "Marshal", "Marshal", ""}, 
-{0x04ef0001, 0x013b0502, "Animal Crossing", "Animal Crossing", "Hazel", "Hazel", ""}, 
-{0x04fa0001, 0x01680502, "Animal Crossing", "Animal Crossing", "Rolf", "Rolf", ""}, 
-{0x04fb0001, 0x01c60502, "Animal Crossing", "Animal Crossing", "Rowan", "Rowan", ""}, 
-{0x04fc0001, 0x02ee0502, "Animal Crossing", "Animal Crossing", "Tybalt", "Tybalt", ""}, 
-{0x04fd0001, 0x007b0502, "Animal Crossing", "Animal Crossing", "Bangle", "Bangle", ""}, 
-{0x04fe0001, 0x00590502, "Animal Crossing", "Animal Crossing", "Leonardo", "Leonardo", ""}, 
-{0x04ff0001, 0x01620502, "Animal Crossing", "Animal Crossing", "Claudia", "Claudia", ""}, 
-{0x05000001, 0x00e70502, "Animal Crossing", "Animal Crossing", "Bianca", "Bianca", ""}, 
-{0x050b0001, 0x00990502, "Animal Crossing", "Animal Crossing", "Chief", "Chief", ""}, 
-{0x050c0001, 0x01c10502, "Animal Crossing", "Animal Crossing", "Lobo", "Lobo", ""}, 
-{0x050d0001, 0x01420502, "Animal Crossing", "Animal Crossing", "Wolfgang", "Wolfgang", ""}, 
-{0x050e0001, 0x00d70502, "Animal Crossing", "Animal Crossing", "Whitney", "Whitney", ""}, 
-{0x050f0001, 0x03140502, "Animal Crossing", "Animal Crossing", "Dobie", "Dobie", ""}, 
-{0x05100001, 0x01070502, "Animal Crossing", "Animal Crossing", "Freya", "Freya", ""}, 
-{0x05110001, 0x01950502, "Animal Crossing", "Animal Crossing", "Fang", "Fang", ""}, 
-{0x05130001, 0x02e70502, "Animal Crossing", "Animal Crossing", "Vivian", "Vivian", ""}, 
-{0x05140001, 0x01530502, "Animal Crossing", "Animal Crossing", "Skye", "Skye", ""}, 
-{0x05150001, 0x005b0502, "Animal Crossing", "Animal Crossing", "Kyle", "Kyle", ""}, 
-{0x05800000, 0x00050002, "Super Smash Bros.", "Star Fox", "Fox", "Fox", ""}, 
-{0x05810000, 0x001c0002, "Super Smash Bros.", "Star Fox", "Falco", "Falco", ""}, 
-{0x05840000, 0x037e0002, "Super Smash Bros.", "Star Fox", "Wolf", "Wolf", ""}, 
-{0x05c00000, 0x00060002, "Super Smash Bros.", "Metroid", "Samus", "Samus", ""}, 
-{0x05c00000, 0x03651302, "Metroid", "Metroid", "Samus", "Samus Aran", ""}, 
-{0x05c00000, 0x04121302, "Metroid", "Metroid", "Samus", "Samus - Metroid Dread", ""}, 
-{0x05c00100, 0x001d0002, "Super Smash Bros.", "Metroid", "Samus", "Zero Suit Samus", ""}, 
-{0x05c10000, 0x03661302, "Metroid", "Metroid", "Metroid", "Metroid", ""}, 
-{0x05c20000, 0x037f0002, "Super Smash Bros.", "Metroid", "Ridley", "Ridley", ""}, 
-{0x05c30000, 0x03800002, "Super Smash Bros.", "Metroid", "Dark Samus", "Dark Samus", ""}, 
-{0x05c40000, 0x04131302, "Metroid", "Metroid", "E.M.M.I.", "E.M.M.I.", ""}, 
-{0x06000000, 0x00120002, "Super Smash Bros.", "F-Zero", "Captain Falcon", "Captain Falcon", ""}, 
-{0x06400100, 0x001e0002, "Super Smash Bros.", "Pikmin", "Olimar", "Olimar", ""}, 
-{0x06420000, 0x035f1102, "Pikmin", "Pikmin", "Pikmin", "Pikmin", ""}, 
-{0x06c00000, 0x000f0002, "Super Smash Bros.", "Punch Out", "Little Mac", "Little Mac", ""}, 
-{0x07000000, 0x00070002, "Super Smash Bros.", "Wii Fit", "Wii Fit Trainer", "Wii Fit Trainer", ""}, 
-{0x07400000, 0x00100002, "Super Smash Bros.", "Kid Icarus", "Pit", "Pit", ""}, 
-{0x07410000, 0x00200002, "Super Smash Bros.", "Kid Icarus", "Dark Pit", "Dark Pit", ""}, 
-{0x07420000, 0x001f0002, "Super Smash Bros.", "Kid Icarus", "Palutena", "Palutena", ""}, 
-{0x07800000, 0x002d0002, "Super Smash Bros.", "Classic Nintendo", "Mr. G&W", "Mr. Game & Watch", ""}, 
-{0x07810000, 0x002e0002, "Super Smash Bros.", "Classic Nintendo", "R.O.B.", "R.O.B. - Famicom", ""}, 
-{0x07810000, 0x00330002, "Super Smash Bros.", "Classic Nintendo", "R.O.B.", "R.O.B. - NES", ""}, 
-{0x07820000, 0x002f0002, "Super Smash Bros.", "Classic Nintendo", "Duck Hunt", "Duck Hunt", ""}, 
-{0x078f0000, 0x03810002, "Super Smash Bros.", "Classic Nintendo", "Ice Climbers", "Ice Climbers", ""}, 
-{0x07c00000, 0x00210002, "Super Smash Bros.", "Mii", "Mii", "Mii Brawler", ""}, 
-{0x07c00100, 0x00220002, "Super Smash Bros.", "Mii", "Mii", "Mii Swordfighter", ""}, 
-{0x07c00200, 0x00230002, "Super Smash Bros.", "Mii", "Mii", "Mii Gunner", ""}, 
-{0x08000100, 0x003e0402, "Splatoon", "Splatoon", "Inkling", "Inkling Girl", ""}, 
-{0x08000100, 0x025f0402, "Splatoon", "Splatoon", "Inkling", "Inkling Girl - Lime Green", ""}, 
-{0x08000100, 0x03690402, "Splatoon", "Splatoon", "Inkling", "Inkling Girl - Neon Pink", ""}, 
-{0x08000100, 0x03820002, "Super Smash Bros.", "Splatoon", "Inkling", "Inkling", ""}, 
-{0x08000100, 0x04150402, "Splatoon", "Splatoon", "Inkling", "Inkling - Yellow", ""}, 
-{0x08000200, 0x003f0402, "Splatoon", "Splatoon", "Inkling", "Inkling Boy", ""}, 
-{0x08000200, 0x02600402, "Splatoon", "Splatoon", "Inkling", "Inkling Boy - Purple", ""}, 
-{0x08000200, 0x036a0402, "Splatoon", "Splatoon", "Inkling", "Inkling Boy - Neon Green", ""}, 
-{0x08000300, 0x00400402, "Splatoon", "Splatoon", "Inkling", "Inkling Squid", ""}, 
-{0x08000300, 0x02610402, "Splatoon", "Splatoon", "Inkling", "Inkling Squid - Orange", ""}, 
-{0x08000300, 0x036b0402, "Splatoon", "Splatoon", "Inkling", "Inkling Squid - Neon Purple", ""}, 
-{0x08010000, 0x025d0402, "Splatoon", "Splatoon", "Callie", "Callie", ""}, 
-{0x08020000, 0x025e0402, "Splatoon", "Splatoon", "Marie", "Marie", ""}, 
-{0x08030000, 0x03760402, "Splatoon", "Splatoon", "Pearl", "Pearl", ""}, 
-{0x08040000, 0x03770402, "Splatoon", "Splatoon", "Marina", "Marina", ""}, 
-{0x08050100, 0x038e0402, "Splatoon", "Splatoon", "Octoling", "Octoling Girl", ""}, 
-{0x08050200, 0x038f0402, "Splatoon", "Splatoon", "Octoling", "Octoling Boy", ""}, 
-{0x08050200, 0x041b0402, "Splatoon", "Splatoon", "Octoling", "Octoling - Blue", ""}, 
-{0x08050300, 0x03900402, "Splatoon", "Splatoon", "Octoling", "Octoling Octopus", ""}, 
-{0x08060100, 0x041c0402, "Splatoon", "Splatoon", "Smallfry", "Smallfry", ""}, 
-{0x09c00101, 0x02690e02, "Mario Sports Superstars", "Mario Sports Superstars", "Mario", "Mario - Soccer", ""}, 
-{0x09c00201, 0x026a0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Mario", "Mario - Baseball", ""}, 
-{0x09c00301, 0x026b0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Mario", "Mario - Tennis", ""}, 
-{0x09c00401, 0x026c0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Mario", "Mario - Golf", ""}, 
-{0x09c00501, 0x026d0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Mario", "Mario - Horse Racing", ""}, 
-{0x09c10101, 0x026e0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Luigi", "Luigi - Soccer", ""}, 
-{0x09c10201, 0x026f0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Luigi", "Luigi - Baseball", ""}, 
-{0x09c10301, 0x02700e02, "Mario Sports Superstars", "Mario Sports Superstars", "Luigi", "Luigi - Tennis", ""}, 
-{0x09c10401, 0x02710e02, "Mario Sports Superstars", "Mario Sports Superstars", "Luigi", "Luigi - Golf", ""}, 
-{0x09c10501, 0x02720e02, "Mario Sports Superstars", "Mario Sports Superstars", "Luigi", "Luigi - Horse Racing", ""}, 
-{0x09c20101, 0x02730e02, "Mario Sports Superstars", "Mario Sports Superstars", "Peach", "Peach - Soccer", ""}, 
-{0x09c20201, 0x02740e02, "Mario Sports Superstars", "Mario Sports Superstars", "Peach", "Peach - Baseball", ""}, 
-{0x09c20301, 0x02750e02, "Mario Sports Superstars", "Mario Sports Superstars", "Peach", "Peach - Tennis", ""}, 
-{0x09c20401, 0x02760e02, "Mario Sports Superstars", "Mario Sports Superstars", "Peach", "Peach - Golf", ""}, 
-{0x09c20501, 0x02770e02, "Mario Sports Superstars", "Mario Sports Superstars", "Peach", "Peach - Horse Racing", ""}, 
-{0x09c30101, 0x02780e02, "Mario Sports Superstars", "Mario Sports Superstars", "Daisy", "Daisy - Soccer", ""}, 
-{0x09c30201, 0x02790e02, "Mario Sports Superstars", "Mario Sports Superstars", "Daisy", "Daisy - Baseball", ""}, 
-{0x09c30301, 0x027a0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Daisy", "Daisy - Tennis", ""}, 
-{0x09c30401, 0x027b0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Daisy", "Daisy - Golf", ""}, 
-{0x09c30501, 0x027c0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Daisy", "Daisy - Horse Racing", ""}, 
-{0x09c40101, 0x027d0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Yoshi", "Yoshi - Soccer", ""}, 
-{0x09c40201, 0x027e0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Yoshi", "Yoshi - Baseball", ""}, 
-{0x09c40301, 0x027f0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Yoshi", "Yoshi - Tennis", ""}, 
-{0x09c40401, 0x02800e02, "Mario Sports Superstars", "Mario Sports Superstars", "Yoshi", "Yoshi - Golf", ""}, 
-{0x09c40501, 0x02810e02, "Mario Sports Superstars", "Mario Sports Superstars", "Yoshi", "Yoshi - Horse Racing", ""}, 
-{0x09c50101, 0x02820e02, "Mario Sports Superstars", "Mario Sports Superstars", "Wario", "Wario - Soccer", ""}, 
-{0x09c50201, 0x02830e02, "Mario Sports Superstars", "Mario Sports Superstars", "Wario", "Wario - Baseball", ""}, 
-{0x09c50301, 0x02840e02, "Mario Sports Superstars", "Mario Sports Superstars", "Wario", "Wario - Tennis", ""}, 
-{0x09c50401, 0x02850e02, "Mario Sports Superstars", "Mario Sports Superstars", "Wario", "Wario - Golf", ""}, 
-{0x09c50501, 0x02860e02, "Mario Sports Superstars", "Mario Sports Superstars", "Wario", "Wario - Horse Racing", ""}, 
-{0x09c60101, 0x02870e02, "Mario Sports Superstars", "Mario Sports Superstars", "Waluigi", "Waluigi - Soccer", ""}, 
-{0x09c60201, 0x02880e02, "Mario Sports Superstars", "Mario Sports Superstars", "Waluigi", "Waluigi - Baseball", ""}, 
-{0x09c60301, 0x02890e02, "Mario Sports Superstars", "Mario Sports Superstars", "Waluigi", "Waluigi - Tennis", ""}, 
-{0x09c60401, 0x028a0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Waluigi", "Waluigi - Golf", ""}, 
-{0x09c60501, 0x028b0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Waluigi", "Waluigi - Horse Racing", ""}, 
-{0x09c70101, 0x028c0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Donkey Kong", "Donkey Kong - Soccer", ""}, 
-{0x09c70201, 0x028d0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Donkey Kong", "Donkey Kong - Baseball", ""}, 
-{0x09c70301, 0x028e0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Donkey Kong", "Donkey Kong - Tennis", ""}, 
-{0x09c70401, 0x028f0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Donkey Kong", "Donkey Kong - Golf", ""}, 
-{0x09c70501, 0x02900e02, "Mario Sports Superstars", "Mario Sports Superstars", "Donkey Kong", "Donkey Kong - Horse Racing", ""}, 
-{0x09c80101, 0x02910e02, "Mario Sports Superstars", "Mario Sports Superstars", "Diddy Kong", "Diddy Kong - Soccer", ""}, 
-{0x09c80201, 0x02920e02, "Mario Sports Superstars", "Mario Sports Superstars", "Diddy Kong", "Diddy Kong - Baseball", ""}, 
-{0x09c80301, 0x02930e02, "Mario Sports Superstars", "Mario Sports Superstars", "Diddy Kong", "Diddy Kong - Tennis", ""}, 
-{0x09c80401, 0x02940e02, "Mario Sports Superstars", "Mario Sports Superstars", "Diddy Kong", "Diddy Kong - Golf", ""}, 
-{0x09c80501, 0x02950e02, "Mario Sports Superstars", "Mario Sports Superstars", "Diddy Kong", "Diddy Kong - Horse Racing", ""}, 
-{0x09c90101, 0x02960e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser", "Bowser - Soccer", ""}, 
-{0x09c90201, 0x02970e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser", "Bowser - Baseball", ""}, 
-{0x09c90301, 0x02980e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser", "Bowser - Tennis", ""}, 
-{0x09c90401, 0x02990e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser", "Bowser - Golf", ""}, 
-{0x09c90501, 0x029a0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser", "Bowser - Horse Racing", ""}, 
-{0x09ca0101, 0x029b0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser Jr.", "Bowser Jr. - Soccer", ""}, 
-{0x09ca0201, 0x029c0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser Jr.", "Bowser Jr. - Baseball", ""}, 
-{0x09ca0301, 0x029d0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser Jr.", "Bowser Jr. - Tennis", ""}, 
-{0x09ca0401, 0x029e0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser Jr.", "Bowser Jr. - Golf", ""}, 
-{0x09ca0501, 0x029f0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Bowser Jr.", "Bowser Jr. - Horse Racing", ""}, 
-{0x09cb0101, 0x02a00e02, "Mario Sports Superstars", "Mario Sports Superstars", "Boo", "Boo - Soccer", ""}, 
-{0x09cb0201, 0x02a10e02, "Mario Sports Superstars", "Mario Sports Superstars", "Boo", "Boo - Baseball", ""}, 
-{0x09cb0301, 0x02a20e02, "Mario Sports Superstars", "Mario Sports Superstars", "Boo", "Boo - Tennis", ""}, 
-{0x09cb0401, 0x02a30e02, "Mario Sports Superstars", "Mario Sports Superstars", "Boo", "Boo - Golf", ""}, 
-{0x09cb0501, 0x02a40e02, "Mario Sports Superstars", "Mario Sports Superstars", "Boo", "Boo - Horse Racing", ""}, 
-{0x09cc0101, 0x02a50e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Mario", "Baby Mario - Soccer", ""}, 
-{0x09cc0201, 0x02a60e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Mario", "Baby Mario - Baseball", ""}, 
-{0x09cc0301, 0x02a70e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Mario", "Baby Mario - Tennis", ""}, 
-{0x09cc0401, 0x02a80e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Mario", "Baby Mario - Golf", ""}, 
-{0x09cc0501, 0x02a90e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Mario", "Baby Mario - Horse Racing", ""}, 
-{0x09cd0101, 0x02aa0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Luigi", "Baby Luigi - Soccer", ""}, 
-{0x09cd0201, 0x02ab0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Luigi", "Baby Luigi - Baseball", ""}, 
-{0x09cd0301, 0x02ac0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Luigi", "Baby Luigi - Tennis", ""}, 
-{0x09cd0401, 0x02ad0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Luigi", "Baby Luigi - Golf", ""}, 
-{0x09cd0501, 0x02ae0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Baby Luigi", "Baby Luigi - Horse Racing", ""}, 
-{0x09ce0101, 0x02af0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Birdo", "Birdo - Soccer", ""}, 
-{0x09ce0201, 0x02b00e02, "Mario Sports Superstars", "Mario Sports Superstars", "Birdo", "Birdo - Baseball", ""}, 
-{0x09ce0301, 0x02b10e02, "Mario Sports Superstars", "Mario Sports Superstars", "Birdo", "Birdo - Tennis", ""}, 
-{0x09ce0401, 0x02b20e02, "Mario Sports Superstars", "Mario Sports Superstars", "Birdo", "Birdo - Golf", ""}, 
-{0x09ce0501, 0x02b30e02, "Mario Sports Superstars", "Mario Sports Superstars", "Birdo", "Birdo - Horse Racing", ""}, 
-{0x09cf0101, 0x02b40e02, "Mario Sports Superstars", "Mario Sports Superstars", "Rosalina", "Rosalina - Soccer", ""}, 
-{0x09cf0201, 0x02b50e02, "Mario Sports Superstars", "Mario Sports Superstars", "Rosalina", "Rosalina - Baseball", ""}, 
-{0x09cf0301, 0x02b60e02, "Mario Sports Superstars", "Mario Sports Superstars", "Rosalina", "Rosalina - Tennis", ""}, 
-{0x09cf0401, 0x02b70e02, "Mario Sports Superstars", "Mario Sports Superstars", "Rosalina", "Rosalina - Golf", ""}, 
-{0x09cf0501, 0x02b80e02, "Mario Sports Superstars", "Mario Sports Superstars", "Rosalina", "Rosalina - Horse Racing", ""}, 
-{0x09d00101, 0x02b90e02, "Mario Sports Superstars", "Mario Sports Superstars", "Metal Mario", "Metal Mario - Soccer", ""}, 
-{0x09d00201, 0x02ba0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Metal Mario", "Metal Mario - Baseball", ""}, 
-{0x09d00301, 0x02bb0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Metal Mario", "Metal Mario - Tennis", ""}, 
-{0x09d00401, 0x02bc0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Metal Mario", "Metal Mario - Golf", ""}, 
-{0x09d00501, 0x02bd0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Metal Mario", "Metal Mario - Horse Racing", ""}, 
-{0x09d10101, 0x02be0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Pink Gold Peach", "Pink Gold Peach - Soccer", ""}, 
-{0x09d10201, 0x02bf0e02, "Mario Sports Superstars", "Mario Sports Superstars", "Pink Gold Peach", "Pink Gold Peach - Baseball", ""}, 
-{0x09d10301, 0x02c00e02, "Mario Sports Superstars", "Mario Sports Superstars", "Pink Gold Peach", "Pink Gold Peach - Tennis", ""}, 
-{0x09d10401, 0x02c10e02, "Mario Sports Superstars", "Mario Sports Superstars", "Pink Gold Peach", "Pink Gold Peach - Golf", ""}, 
-{0x09d10501, 0x02c20e02, "Mario Sports Superstars", "Mario Sports Superstars", "Pink Gold Peach", "Pink Gold Peach - Horse Racing", ""}, 
-{0x0a000001, 0x03ab0502, "Animal Crossing", "Animal Crossing", "Orville", "Orville", ""}, 
-{0x0a010001, 0x03ac0502, "Animal Crossing", "Animal Crossing", "Wilbur", "Wilbur", ""}, 
-{0x0a020001, 0x03b30502, "Animal Crossing", "Animal Crossing", "C.J.", "C.J.", ""}, 
-{0x0a030001, 0x03b40502, "Animal Crossing", "Animal Crossing", "Flick", "Flick", ""}, 
-{0x0a040001, 0x03b50502, "Animal Crossing", "Animal Crossing", "Daisy Mae", "Daisy Mae", ""}, 
-{0x0a050001, 0x03b80502, "Animal Crossing", "Animal Crossing", "Harvey", "Harvey", ""}, 
-{0x0a060001, 0x03ba0502, "Animal Crossing", "Animal Crossing", "Wisp", "Wisp", ""}, 
-{0x0a070001, 0x03bc0502, "Animal Crossing", "Animal Crossing", "Niko", "Niko", ""}, 
-{0x0a080001, 0x03bd0502, "Animal Crossing", "Animal Crossing", "Wardell", "Wardell", ""}, 
-{0x0a090001, 0x03c00502, "Animal Crossing", "Animal Crossing", "Sherb", "Sherb", ""}, 
-{0x0a0a0001, 0x03c10502, "Animal Crossing", "Animal Crossing", "Megan", "Megan", ""}, 
-{0x0a0b0001, 0x03c20502, "Animal Crossing", "Animal Crossing", "Dom", "Dom", ""}, 
-{0x0a0c0001, 0x03c30502, "Animal Crossing", "Animal Crossing", "Audie", "Audie", ""}, 
-{0x0a0d0001, 0x03c40502, "Animal Crossing", "Animal Crossing", "Cyd", "Cyd", ""}, 
-{0x0a0e0001, 0x03c50502, "Animal Crossing", "Animal Crossing", "Judy", "Judy", ""}, 
-{0x0a0f0001, 0x03c60502, "Animal Crossing", "Animal Crossing", "Raymond", "Raymond", ""}, 
-{0x0a100001, 0x03c70502, "Animal Crossing", "Animal Crossing", "Reneigh", "Reneigh", ""}, 
-{0x0a110001, 0x03c80502, "Animal Crossing", "Animal Crossing", "Sasha", "Sasha", ""}, 
-{0x0a120001, 0x03c90502, "Animal Crossing", "Animal Crossing", "Ione", "Ione", ""}, 
-{0x0a130001, 0x03ca0502, "Animal Crossing", "Animal Crossing", "Tiansheng", "Tiansheng", ""}, 
-{0x0a140001, 0x03cb0502, "Animal Crossing", "Animal Crossing", "Shino", "Shino", ""}, 
-{0x0a150001, 0x03cc0502, "Animal Crossing", "Animal Crossing", "Marlo", "Marlo", ""}, 
-{0x0a160001, 0x03cd0502, "Animal Crossing", "Animal Crossing", "Petri", "Petri", ""}, 
-{0x0a170001, 0x03ce0502, "Animal Crossing", "Animal Crossing", "Cephalobot", "Cephalobot", ""}, 
-{0x0a180001, 0x03cf0502, "Animal Crossing", "Animal Crossing", "Quinn", "Quinn", ""}, 
-{0x0a190001, 0x03d00502, "Animal Crossing", "Animal Crossing", "Chabwick", "Chabwick", ""}, 
-{0x0a1a0001, 0x03d10502, "Animal Crossing", "Animal Crossing", "Zoe", "Zoe", ""}, 
-{0x0a1b0001, 0x03d20502, "Animal Crossing", "Animal Crossing", "Ace", "Ace", ""}, 
-{0x0a1c0001, 0x03d30502, "Animal Crossing", "Animal Crossing", "Rio", "Rio", ""}, 
-{0x0a1d0001, 0x03d40502, "Animal Crossing", "Animal Crossing", "Frett", "Frett", ""}, 
-{0x0a1e0001, 0x03d50502, "Animal Crossing", "Animal Crossing", "Azalea", "Azalea", ""}, 
-{0x0a1f0001, 0x03d60502, "Animal Crossing", "Animal Crossing", "Roswell", "Roswell", ""}, 
-{0x0a200001, 0x03d70502, "Animal Crossing", "Animal Crossing", "Faith", "Faith", ""}, 
-{0x0a400000, 0x041d0002, "Super Smash Bros.", "ARMS", "Min Min", "Min Min", ""}, 
-{0x19020000, 0x03830002, "Super Smash Bros.", "Pokemon", "Ivysaur", "Ivysaur", ""}, 
-{0x19060000, 0x00240002, "Super Smash Bros.", "Pokemon", "Charizard", "Charizard", ""}, 
-{0x19070000, 0x03840002, "Super Smash Bros.", "Pokemon", "Squirtle", "Squirtle", ""}, 
-{0x19190000, 0x00090002, "Super Smash Bros.", "Pokemon", "Pikachu", "Pikachu", ""}, 
-{0x19270000, 0x00260002, "Super Smash Bros.", "Pokemon", "Jigglypuff", "Jigglypuff", ""}, 
-{0x19960000, 0x023d0002, "Super Smash Bros.", "Pokemon", "Mewtwo", "Mewtwo", ""}, 
-{0x19ac0000, 0x03850002, "Super Smash Bros.", "Pokemon", "Pichu", "Pichu", ""}, 
-{0x1ac00000, 0x00110002, "Super Smash Bros.", "Pokemon", "Lucario", "Lucario", ""}, 
-{0x1b920000, 0x00250002, "Super Smash Bros.", "Pokemon", "Greninja", "Greninja", ""}, 
-{0x1bd70000, 0x03860002, "Super Smash Bros.", "Pokemon", "Incineroar", "Incineroar", ""}, 
-{0x1d000001, 0x025c0d02, "Pokemon", "Pokemon", "Shadow Mewtwo", "Shadow Mewtwo", ""}, 
-{0x1d010000, 0x03750d02, "Pokemon", "Pokemon", "Detective Pikachu", "Detective Pikachu", ""}, 
-{0x1d400000, 0x03870002, "Super Smash Bros.", "Pokemon", "Pokemon Trainer", "Pokemon Trainer", ""}, 
-{0x1f000000, 0x000a0002, "Super Smash Bros.", "Kirby", "Kirby", "Kirby", ""}, 
-{0x1f000000, 0x02540c02, "Kirby", "Kirby", "Kirby", "Kirby", ""}, 
-{0x1f010000, 0x00270002, "Super Smash Bros.", "Kirby", "Meta Knight", "Meta Knight", ""}, 
-{0x1f010000, 0x02550c02, "Kirby", "Kirby", "Meta Knight", "Meta Knight", ""}, 
-{0x1f020000, 0x00280002, "Super Smash Bros.", "Kirby", "King Dedede", "King Dedede", ""}, 
-{0x1f020000, 0x02560c02, "Kirby", "Kirby", "King Dedede", "King Dedede", ""}, 
-{0x1f030000, 0x02570c02, "Kirby", "Kirby", "Waddle Dee", "Waddle Dee", ""}, 
-{0x1f400000, 0x035e1002, "BoxBoy!", "BoxBoy!", "Qbby", "Qbby", ""}, 
-{0x21000000, 0x000b0002, "Super Smash Bros.", "Fire Emblem", "Marth", "Marth", ""}, 
-{0x21010000, 0x00180002, "Super Smash Bros.", "Fire Emblem", "Ike", "Ike", ""}, 
-{0x21020000, 0x00290002, "Super Smash Bros.", "Fire Emblem", "Lucina", "Lucina", ""}, 
-{0x21030000, 0x002a0002, "Super Smash Bros.", "Fire Emblem", "Robin", "Robin", ""}, 
-{0x21040000, 0x02520002, "Super Smash Bros.", "Fire Emblem", "Roy", "Roy", ""}, 
-{0x21050000, 0x025a0002, "Super Smash Bros.", "Fire Emblem", "Corrin", "Corrin", ""}, 
-{0x21050100, 0x03630002, "Super Smash Bros.", "Fire Emblem", "Corrin", "Corrin - Player 2", ""}, 
-{0x21060000, 0x03601202, "Fire Emblem", "Fire Emblem", "Alm", "Alm", ""}, 
-{0x21070000, 0x03611202, "Fire Emblem", "Fire Emblem", "Celica", "Celica", ""}, 
-{0x21080000, 0x036f1202, "Fire Emblem", "Fire Emblem", "Chrom", "Chrom", ""}, 
-{0x21080000, 0x03880002, "Super Smash Bros.", "Fire Emblem", "Chrom", "Chrom", ""}, 
-{0x21090000, 0x03701202, "Fire Emblem", "Fire Emblem", "Tiki", "Tiki", ""}, 
-{0x210b0000, 0x03a50002, "Super Smash Bros.", "Fire Emblem", "Byleth", "Byleth", ""}, 
-{0x22400000, 0x002b0002, "Super Smash Bros.", "Xenoblade", "Shulk", "Shulk", ""}, 
-{0x22800000, 0x002c0002, "Super Smash Bros.", "Earthbound", "Ness", "Ness", ""}, 
-{0x22810000, 0x02510002, "Super Smash Bros.", "Earthbound", "Lucas", "Lucas", ""}, 
-{0x22c00000, 0x003a0202, "Chibi-Robo!", "Chibi Robo", "Chibi-Robo", "Chibi Robo", ""}, 
-{0x32000000, 0x00300002, "Super Smash Bros.", "Sonic", "Sonic", "Sonic", ""}, 
-{0x32400000, 0x025b0002, "Super Smash Bros.", "Bayonetta", "Bayonetta", "Bayonetta", ""}, 
-{0x32400100, 0x03640002, "Super Smash Bros.", "Bayonetta", "Bayonetta", "Bayonetta - Player 2", ""}, 
-{0x33400000, 0x00320002, "Super Smash Bros.", "Pac-man", "PAC-MAN", "Pac-Man", ""}, 
-{0x33800000, 0x03781402, "Others", "Dark Souls", "Solaire of Astora", "Solaire of Astora", ""}, 
-{0x34800000, 0x00310002, "Super Smash Bros.", "Megaman", "Mega Man", "Mega Man", ""}, 
-{0x34800000, 0x02580002, "Super Smash Bros.", "Megaman", "Mega Man", "Mega Man - Gold Edition", ""}, 
-{0x34800000, 0x03791502, "Mega Man", "Megaman", "Mega Man", "Mega Man", ""}, 
-{0x34c00000, 0x02530002, "Super Smash Bros.", "Street fighter", "Ryu", "Ryu", ""}, 
-{0x34c10000, 0x03890002, "Super Smash Bros.", "Street fighter", "Ken", "Ken", ""}, 
-{0x35000100, 0x02e10f02, "Monster Hunter", "Monster Hunter", "One-Eyed Rathalos", "One-Eyed Rathalos and Rider - Male", ""}, 
-{0x35000200, 0x02e20f02, "Monster Hunter", "Monster Hunter", "One-Eyed Rathalos", "One-Eyed Rathalos and Rider - Female", ""}, 
-{0x35010000, 0x02e30f02, "Monster Hunter", "Monster Hunter", "Nabiru", "Nabiru", ""}, 
-{0x35020100, 0x02e40f02, "Monster Hunter", "Monster Hunter", "Rathian", "Rathian and Cheval", ""}, 
-{0x35030100, 0x02e50f02, "Monster Hunter", "Monster Hunter", "Barioth", "Barioth and Ayuria", ""}, 
-{0x35040100, 0x02e60f02, "Monster Hunter", "Monster Hunter", "Qurupeco", "Qurupeco and Dan", ""}, 
-{0x35050000, 0x040c0f02, "Monster Hunter", "Monster Hunter", "Razewing Ratha", "Razewing Ratha", ""}, 
-{0x35060000, 0x040d0f02, "Monster Hunter", "Monster Hunter", "Ena", "Ena", ""}, 
-{0x35070000, 0x040e0f02, "Monster Hunter", "Monster Hunter", "Tsukino", "Tsukino", ""}, 
-{0x35080000, 0x040f1802, "Monster Hunter Rise", "Monster Hunter", "Magnamalo", "Magnamalo", ""}, 
-{0x35090000, 0x04101802, "Monster Hunter Rise", "Monster Hunter", "Palico", "Palico", ""}, 
-{0x35090100, 0x042b1802, "Monster Hunter Rise", "Monster Hunter", "Palico", "Palico", ""}, 
-{0x350a0000, 0x04111802, "Monster Hunter Rise", "Monster Hunter", "Palamute", "Palamute", ""}, 
-{0x350a0100, 0x042c1802, "Monster Hunter Rise", "Monster Hunter", "Palamute", "Palamute", ""}, 
-{0x350b0000, 0x042d1802, "Monster Hunter Rise", "Monster Hunter", "Malzeno", "Malzeno", ""}, 
-{0x35c00000, 0x02500a02, "Shovel Knight", "Shovel Knight", "Shovel Knight", "Shovel Knight", ""}, 
-{0x35c00000, 0x03920a02, "Shovel Knight", "Shovel Knight", "Shovel Knight", "Shovel Knight - Gold Edition", ""}, 
-{0x35c10000, 0x036c0a02, "Shovel Knight", "Shovel Knight", "Plague Knight", "Plague Knight", ""}, 
-{0x35c20000, 0x036d0a02, "Shovel Knight", "Shovel Knight", "Specter Knight", "Specter Knight", ""}, 
-{0x35c30000, 0x036e0a02, "Shovel Knight", "Shovel Knight", "King Knight", "King Knight", ""}, 
-{0x36000000, 0x02590002, "Super Smash Bros.", "Final Fantasy", "Cloud Strife", "Cloud", ""}, 
-{0x36000100, 0x03620002, "Super Smash Bros.", "Final Fantasy", "Cloud Strife", "Cloud - Player 2", ""}, 
-{0x36400000, 0x03a20002, "Super Smash Bros.", "Dragon Quest", "Hero", "Hero", ""}, 
-{0x37400001, 0x03741402, "Others", "Kellogs", "Mario Cereal", "Super Mario Cereal", ""}, 
-{0x37800000, 0x038a0002, "Super Smash Bros.", "Metal Gear Solid", "Snake", "Snake", ""}, 
-{0x37c00000, 0x038b0002, "Super Smash Bros.", "Castlevania", "Simon", "Simon", ""}, 
-{0x37c10000, 0x038c0002, "Super Smash Bros.", "Castlevania", "Richter", "Richter", ""}, 
-{0x38000001, 0x03931702, "Power Pros", "Power Pros", "Pawapuro", "Pawapuro", ""}, 
-{0x38010001, 0x03941702, "Power Pros", "Power Pros", "Ikari", "Ikari", ""}, 
-{0x38020001, 0x03951702, "Power Pros", "Power Pros", "Daijobu", "Daijobu", ""}, 
-{0x38030001, 0x03961702, "Power Pros", "Power Pros", "Hayakawa", "Hayakawa", ""}, 
-{0x38040001, 0x03971702, "Power Pros", "Power Pros", "Yabe", "Yabe", ""}, 
-{0x38050001, 0x03981702, "Power Pros", "Power Pros", "Ganda", "Ganda", ""}, 
-{0x38400001, 0x04241902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Yuga Ohdo", "Yuga Ohdo", ""}, 
-{0x38410001, 0x04251902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Tatsuhisa “Luke” Kamijō", "Tatsuhisa “Luke” Kamijō", ""}, 
-{0x38420001, 0x04261902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Gakuto Sōgetsu", "Gakuto Sōgetsu", ""}, 
-{0x38430001, 0x04271902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Romin Kirishima", "Romin Kirishima", ""}, 
-{0x38440001, 0x04281902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Roa Kirishima", "Roa Kirishima", ""}, 
-{0x38450001, 0x04291902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Nail Saionji", "Nail Saionji", ""}, 
-{0x38460001, 0x042a1902, "Yu-Gi-Oh!", "Yu-Gi-Oh!", "Asana Mutsuba", "Asana Mutsuba", ""}, 
-{0x38c00000, 0x03911602, "Diablo", "Diablo", "Loot Goblin", "Loot Goblin", ""}, 
-{0x3a000000, 0x03a10002, "Super Smash Bros.", "Persona", "Joker", "Joker", ""}, 
-{0x3b400000, 0x03a30002, "Super Smash Bros.", "Banjo Kazooie", "Banjo", "Banjo & Kazooie", ""}, 
-{0x3c800000, 0x03a40002, "Super Smash Bros.", "Fatal Fury", "Terry", "Terry", ""}, 
-{0x3dc00000, 0x04220002, "Super Smash Bros.", "Minecraft", "Steve", "Steve", ""}, 
-{0x3dc10000, 0x04230002, "Super Smash Bros.", "Minecraft", "Alex", "Alex", ""}, 
-{0x33c00000, 0x04200002, "Super Smash Bros.", "Tekken", "Kazuya", "Kazuya", ""}, 
-{0x36010000, 0x04210002, "Super Smash Bros.", "Final Fantasy", "Sephiroth", "Sephiroth", ""}, 
-{0x00000003, 0x0430ff02, "Super Nintendo World", "Super Mario", "Mario", "Golden - Power Up Band", ""}, 
-{0x01000000, 0x04180902, "Legend Of Zelda", "The Legend of Zelda", "Link", "Link - Tears of the Kingdom", ""}, 
-{0, 0, 0, 0, 0, 0}
-};
