--- fw/application/src/amiibo_data.h
@@ -1,17 +1,0 @@
-/*
- * amiibo_data.h
- *
- *  Created on: 2021年9月13日
- *      Author: solos
- */
-
-#ifndef AMIIBO_DATA_H_
-#define AMIIBO_DATA_H_
-
-#include <stdint.h>
-#include <stdbool.h>
-
-uint32_t to_little_endian_int32(uint8_t* data);
-
-
-#endif /* AMIIBO_DATA_H_ */
