+++ fw/application/src/app/amiibo/scene/amiibo_scene_amiibo_detail_menu.c
@@ -1,4 +1,3 @@
-#include "amiibo_data.h"
 #include "amiibo_helper.h"
 #include "amiibo_scene.h"
 #include "app_amiibo.h"
